Composite models

The model algebra defined by the AbstractSpectralModelKind yields instances of CompositeSpectralModel, nested to various degrees. These composite models are designed to make as much information about the spectral model available at compile-time, such that rich and optimized generated functions may be assembled purely from the Julia types (see Why & how).

SpectralFitting.CompositeSpectralModel — Type

CompositeSpectralModel{M1,M2,O} <: AbstractSpectralModel
CompositeSpectralModel(left_model, right_model, op::AbstractCompositeOperator)

Type resulting from operations combining any number of AbstractSpectralModel via the model algebra defined from AbstractSpectralModelKind.

Each operation binary operation in the model algebra is encoded in the parametric types of the CompositeSpectralModel, where the operation is given by an AbstractCompositeOperator. Composite models adopt the model kind of the right model, i.e. M2, and obey the model algebra accordingly.

Composite models very rarely need to be constructed directly, and are instead obtained by regular model operations.

Example

model = XS_PhotoelectricAbsorption() * (XS_PowerLaw() + XS_BlackBody())
typeof(model) <: CompositeSpectralModel # true

SpectralFitting.AbstractCompositeOperator — Type

abstract type AbstractCompositeOperator

Superype of all composition operators. Used to implement the model algebra of AbstractSpectralModelKind through a trait system.

The Julia symbol corresponding to a given AbstractCompositeOperator may be obtained through operation_symbol.

SpectralFitting.AdditionOperator — Type

AdditionOperator <: AbstractCompositeOperator
AdditionOperator()

Corresponds to the :(+) symbol.

SpectralFitting.MultiplicationOperator — Type

MultiplicationOperator <: AbstractCompositeOperator
MultiplicationOperator()

Corresponds to the :(*) symbol.

SpectralFitting.ConvolutionOperator — Type

ConvolutionOperator <: AbstractCompositeOperator
ConvolutionOperator()

Has no corresponding symbol, since it invokes a function call C(A).

SpectralFitting.operation_symbol — Function

operation_symbol(::AbstractCompositeOperator)
operation_symbol(::Type{<:AbstractCompositeOperator})

Obtain the model symbol from a given AbstractCompositeOperator.