Models

Models exist as three different kinds

SpectralFitting.AbstractSpectralModelKind — Type

abstract type AbstractSpectralModelKind

Abstract type of all model kinds.

SpectralFitting.modelkind — Function

modelkind(M::Type{<:AbstractSpectralModel})

Return the kind of model given by M: either Additive, Multiplicative, or Convolutional.

SpectralFitting.Additive — Type

Additive <: AbstractSpectralModelKind
Additive()

Additive models are effectively the sources of photons, and are the principle building blocks of composite models.

Note

Defining custom additive models requires special care. See Defining new models.

SpectralFitting.Multiplicative — Type

Multiplicative <: AbstractSpectralModelKind
Multiplicative()

Multiplicative models act on Additive models, by element-wise multiplying the flux in each energy bin of the additive model by a different factor.

SpectralFitting.Convolutional — Type

Convolutional <: AbstractSpectralModelKind
Convolutional()

Convolutional models act on the flux generated by Additive models, similar to Multiplicative models, however may convolve kernels through the flux also.

Models wrapped from XSPEC implementations are prefixed with XS_*, whereas pure-Julia models are simply named, e.g. XS_PowerLaw in XSPEC vs PowerLaw in Julia.

The available models are

SpectralFitting.AbstractSpectralModel

SpectralFitting.AbstractSpectralModelKind

SpectralFitting.Additive

SpectralFitting.BlackBody

SpectralFitting.Convolutional

SpectralFitting.Multiplicative

SpectralFitting.PowerLaw

SpectralFitting.SurrogateSpectralModel

SpectralFitting.XS_BlackBody

SpectralFitting.XS_BremsStrahlung

SpectralFitting.XS_CalculateFlux

SpectralFitting.XS_DiskLine

SpectralFitting.XS_Laor

SpectralFitting.XS_PhotoelectricAbsorption

SpectralFitting.XS_PowerLaw

SpectralFitting.XS_WarmAbsorption

SpectralFitting.AbstractSpectralModel — Type

abstract type AbstractSpectralModel

Supertype of all spectral models. Sub-types must implement the following interface

modelkind

SpectralFitting.invoke!

SpectralFitting.invoke! — Function

SpectralFitting.invoke!(flux, energy, M::Type{<:AbstractSpectralModel}, params...)

Used to define the behaviour of models. Should calculate flux of the model and write in-place into flux.

Note

This function should not be called directly. Use invokemodel instead.

Parameters are passed in in-order as defined in the model structure. For example

struct MyModel{F1,F2,F3,...} <: AbstractSpectralModel
    p1::F1
    p2::F2
    p3::F3
    # ...
end

would have the arguments passed to invoke! as

function SpectralFitting.invoke!(flux, energy, ::Type{<:MyModel}, p1, p2, p3, ...)
    # ...
end

The only exception to this are Additive models, where the normalisation parameter K is not passed to invoke!.

Using models

Missing docstring.

Missing docstring for invokemodel. Check Documenter's build log for details.

Defining new models

To define a new Julia model, we need only implement the AbstractSpectralModel interface. As a pedagogical example, consider an implementation of PowerLaw:

# define the model and parameters
@with_kw struct PowerLaw{F1,F2} <: AbstractSpectralModel
    "Normalisation."
    K::F1 = FitParam(1.0)
    "Photon index."
    a::F2 = FitParam(0.5)
end

# tell the package what kind of model this is
modelkind(::Type{<:PowerLaw}) = Additive()

# define how the model acts: 
# all of the parameters are passed, in order, as arguments
function SpectralFitting.invoke!(flux, energy, ::Type{<:PowerLaw}, a)
    α = 1 - a
    α⁻¹ = inv(α)
    SpectralFitting.integrate_over_flux!(flux, energy) do E
        α⁻¹ * E^α
    end
end

Note that each parameter has its own parametric type, here ::F1 and ::F2. This is so that the type of fit parameters can be changed, or alternative number types inserted (see Why & how). The standard parameter type is FitParam, which is what we have used here.

The model is defined to be additive by overloading modelkind.

Note

Additive models must have a normalisation parameter with the symbol K, which is, however, not passed to SpectralFitting.invoke!. Multiplicative and convolutional models have no such requirements.

We have also used the SpectralFitting.integrate_over_flux utility function, which is designed to help us efficiently compute finite-difference integrations of flux, just as XSPEC does.

Note

To add additional XSPEC models, see Wrapping new XSPEC models.

Julia models

SpectralFitting.PowerLaw — Type

XS_PowerLaw(K, a)

K

Normalisation. Default: FitParam(1.0)

a

Photon index. Default: FitParam(0.5)

SpectralFitting.BlackBody — Type

XS_BlackBody(K, T)

K

Normalisation. Default: FitParam(1.0)

kT

Temperature (keV). Default: FitParam(2.0)

XSPEC models

SpectralFitting.XS_PowerLaw — Type

XS_PowerLaw(K, a)

K

Normalisation. Default: FitParam(1.0)

a

Photon index. Default: FitParam(0.5)

SpectralFitting.XS_BlackBody — Type

XS_BlackBody(K, T)

K

Normalisation. Default: FitParam(1.0)

T

Temperature (keV). Default: FitParam(3.0)

SpectralFitting.XS_BremsStrahlung — Type

XS_BremsStrahlung(K, T)

K

Normalisation. Default: FitParam(1.0)

T

Plasma temperature (keV). Default: FitParam(7.0)

SpectralFitting.XS_Laor — Type

XS_Laor(K, lineE, a, inner_r, outer_r, incl)

K

Normalisation. Default: FitParam(1.0)

lineE

Rest frame line energy (keV). Default: FitParam(6.4)

a

Power law dependence of emissitivy. Scales R⁻ᵅ. Default: FrozenFitParam(3.0)

inner_r

Inner radius of the accretion disk (GM/c). Default: FrozenFitParam(1.235)

outer_r

Outer radius of the accretion disk (GM/c). Default: FrozenFitParam(400.0)

incl

Disk inclination angle to line of sight (degrees, 0 is pole on). Default: FitParam(30.0)

SpectralFitting.XS_DiskLine — Type

XS_DiskLine(K, lineE, β, inner_r, outer_r, incl)

K

Normalisation. Default: FitParam(1.0)

lineE

Rest frame line energy (keV). Default: FitParam(6.7)

β

Power law dependence of emissitivy. If < 10, scales Rᵅ. Default: FrozenFitParam(-2.0)

inner_r

Inner radius of the accretion disk (GM/c). Default: FrozenFitParam(10.0)

outer_r

Outer radius of the accretion disk (GM/c). Default: FrozenFitParam(1000.0)

incl

Disk inclination angle to line of sight (degrees, 0 is pole on). Default: FitParam(30.0)

SpectralFitting.XS_PhotoelectricAbsorption — Type

XS_PhotoelectricAbsorption(ηH)

ηH

Equivalent hydrogen column (units of 10²² atoms per cm⁻²). Default: FitParam(1.0)

SpectralFitting.XS_WarmAbsorption — Type

XS_WarmAbsorption(ηH, Ew)

ηH

Equivalent hydrogen column (units of 10²² atoms per cm⁻²). Default: FitParam(1.0)

Ew

Window energy (keV). Default: FitParam(1.0)

SpectralFitting.XS_CalculateFlux — Type

XS_CalculateFlux(E_min, E_max, lg10Flux)

E_min

Minimum energy. Default: FrozenFitParam(0.2)

E_max

Maximum energy. Default: FrozenFitParam(2.0)

lg10Flux

log (base 10) flux in erg / cm^2 / s Default: FitParam(-10.0, lowerlimit = -Inf, upperlimit = 0.0)

Wrapping new XSPEC models

SpectralFitting exports a helpful macro to facilitate wrapping additional XSPEC models.

SpectralFitting.@xspecmodel — Macro

@xspecmodel model_kind, func_name, model

Used to wrap additional XSPEC models, generating the needed AbstractSpectralModel implementation.

Examples

@xspecmodel Additive :C_powerlaw struct XS_PowerLaw{F1,F2}
    "Normalisation."
    K::F1 = FitParam(1.0)
    "Photon index."
    a::F2 = FitParam(0.5)
end

We define a new structure XS_PowerLaw with two parameters, but since the model is Additive, only a single parameter (a) is passed to the XSPEC function. The function we bind to this model is :C_powerlaw from the XSPEC C wrappers.

The macro will, in this case, generate the following functions

modelkind(::Type{<:XS_PowerLaw})
implementation(::Type{<:XS_PowerLaw})
invoke!(::Type{<:XS_PowerLaw})

Other models

Missing docstring.

Missing docstring for CompositeSpectralModel. Check Documenter's build log for details.

SpectralFitting.SurrogateSpectralModel — Type

SurrogateSpectralModel <: AbstractSpectralModel
SurrogateSpectralModel(modelkind, surrogate, params, params_symbols)

Used to wrap a surrogate function into an AbstractSpectralModel.

Example

Creating a surrogate function for XS_PhotoelectricAbsorption using wrap_model_as_objective and make_surrogate_function.

using SpectralFitting
using Surrogates

model = XS_PhotoelectricAbsorption()
objsampler = wrap_model_as_objective(model)

# upper and lower bounds on energy and ηH
lb = (0.1, 1e-3)
ub = (30.0, 30.0)

# build surrogate function
surrogate = make_surrogate_function(model, s_lb, s_ub)

# create surrogate model
sm = SurrogateSpectralModel(
    Multiplicative(),
    surrogate,
    (FitParam(1.0),),
    (:ηH,)
)

SpectralFitting.wrap_model_as_objective — Function

wrap_model_as_objective(model::AbstractSpectralModel; ΔE = 1e-1)
wrap_model_as_objective(M::Type{<:AbstractSpectralModel}; ΔE = 1e-1)

Wrap a spectral model into an objective function for building/optimizing a surrogate model. Returns an anonymous function taking the tuple (E, params...) as the argument, and returning a single flux value.

SpectralFitting.make_surrogate_function — Function

make_surrogate_function(
    model::M,
    lowerbounds::T,
    upperbounds::T;
    optimization_samples = 200,
    seed_samples = 50,
    S::Type = RadialBasis,
    sample_type = SobolSample(),
    verbose = false,
)

Creates and optimizes a surrogate model of type S for model, using wrap_model_as_objective and optimize_accuracy! for optimization_samples iterations. Model is initially seeded with seed_samples points prior to optimization.

Note

Additive models integrate energies to calculate flux, which surrogate models are currently not capable of. Results for Additive models likely to be inaccurate. This will be patched in a future version.

SpectralFitting.optimize_accuracy! — Function

optimize_accuracy!(
    surr::AbstractSurrogate,
    obj::Function,
    lb,
    ub;
    sample_type::SamplingAlgorithm = SobolSample(),
    maxiters = 200,
    N_truth = 5000,
    verbose = false,
)

Improve accuracy (faithfullness) of the surrogate model in recreating the objective function.

Samples a new space of N_truth points between lb and ub, and calculates the objective function obj at each. Finds the point with largest MSE between surrogate and objective, and adds the point to the surrogate pool. Repeats maxiters times, adding maxiters points to surrogate model.

Optionally print to stdout the MSE and iteration count with verbose = true.

Note that upper- and lower-bounds should be in the form (E, params...), where E is a single energy and params are the model parameters.