Surrogate models

SpectralFitting.SurrogateSpectralModel — Type

SurrogateSpectralModel <: AbstractSpectralModel
SurrogateSpectralModel(modelkind, surrogate, params, params_symbols)

Used to wrap a surrogate function into an AbstractSpectralModel.

Example

Creating a surrogate function for XS_PhotoelectricAbsorption using wrap_model_as_objective and make_surrogate_function:

using SpectralFitting
using Surrogates

model = XS_PhotoelectricAbsorption()
objsampler = wrap_model_as_objective(model)

# upper and lower bounds on energy and ηH
lb = (0.1, 1e-3)
ub = (30.0, 30.0)

# build surrogate function
surrogate = make_surrogate_function(model, lb, ub)

# create surrogate model
sm = SurrogateSpectralModel(
    Multiplicative(),
    surrogate,
    (FitParam(1.0),),
    (:ηH,)
)

SpectralFitting.wrap_model_as_objective — Function

wrap_model_as_objective(model::AbstractSpectralModel; ΔE = 1e-1)
wrap_model_as_objective(M::Type{<:AbstractSpectralModel}; ΔE = 1e-1)

Wrap a spectral model into an objective function for building/optimizing a surrogate model. Returns an anonymous function taking the tuple (E, params...) as the argument, and returning a single flux value.

SpectralFitting.make_surrogate_function — Function

make_surrogate_function(
    model::M,
    lowerbounds::T,
    upperbounds::T;
    optimization_samples = 200,
    seed_samples = 50,
    S::Type = RadialBasis,
    sample_type = SobolSample(),
    verbose = false,
)

Creates and optimizes a surrogate model of type S for model, using wrap_model_as_objective and optimize_accuracy! for optimization_samples iterations. Model is initially seeded with seed_samples points prior to optimization.

Note

Additive models integrate energies to calculate flux, which surrogate models are currently not capable of. Results for Additive models likely to be inaccurate. This will be patched in a future version.

SpectralFitting.optimize_accuracy! — Function

optimize_accuracy!(
    surr::AbstractSurrogate,
    obj::Function,
    lb,
    ub;
    sample_type::SamplingAlgorithm = SobolSample(),
    maxiters = 200,
    N_truth = 5000,
    verbose = false,
)

Improve accuracy (faithfullness) of the surrogate model in recreating the objective function.

Samples a new space of N_truth points between lb and ub, and calculates the objective function obj at each. Finds the point with largest MSE between surrogate and objective, and adds the point to the surrogate pool. Repeats maxiters times, adding maxiters points to surrogate model.

Optionally print to stdout the MSE and iteration count with verbose = true.

Note that upper- and lower-bounds should be in the form (E, params...), where E is a single energy and params are the model parameters.