Models index

Models wrapped from XSPEC implementations are prefixed with XS_*, whereas pure-Julia models are simply named, e.g. XS_PowerLaw in XSPEC vs PowerLaw in Julia.

The available models are

SpectralFitting.BlackBody

SpectralFitting.PowerLaw

SpectralFitting.XS_BlackBody

SpectralFitting.XS_BremsStrahlung

SpectralFitting.XS_CalculateFlux

SpectralFitting.XS_DiskLine

SpectralFitting.XS_Laor

SpectralFitting.XS_PhotoelectricAbsorption

SpectralFitting.XS_PowerLaw

SpectralFitting.XS_WarmAbsorption

Julia models

SpectralFitting.PowerLaw — Type

XS_PowerLaw(K, a)

K

Normalisation. Default: FitParam(1.0)

a

Photon index. Default: FitParam(0.5)

Example

energy = collect(range(0.1, 20.0, 100))
invokeflux(energy, PowerLaw())

                        PowerLaw
       ┌────────────────────────────────────────┐
   0.5 │                                        │
       │:                                       │
       │:                                       │
       │:                                       │
       │:                                       │
       │:                                       │
       │:                                       │
       │ :                                      │
       │ :                                      │
       │  :                                     │
       │   :.                                   │
       │    ':..                                │
       │        ''':......                      │
       │                  ''''''''''''''........│
     0 │                                        │
       └────────────────────────────────────────┘
        0                                     20
                         E (keV)

SpectralFitting.BlackBody — Type

XS_BlackBody(K, T)

K

Normalisation. Default: FitParam(1.0)

kT

Temperature (keV). Default: FitParam(3.0)

Example

energy = collect(range(0.1, 20.0, 100))
invokeflux(energy, BlackBody())

                        BlackBody
       ┌────────────────────────────────────────┐
   0.2 │                                        │
       │                                        │
       │                                        │
       │                                        │
       │                                        │
       │                                        │
       │      .:''':..                          │
       │     :'      '':.                       │
       │   .'           ':.                     │
       │  .:               '..                  │
       │  :                  ':.                │
       │ .'                     ':..            │
       │ :                         ''...        │
       │:                              '''....  │
     0 │:                                    '''│
       └────────────────────────────────────────┘
        0                                     20
                         E (keV)

Defining new models

To define a new Julia model, we need only implement the AbstractSpectralModel interface. As a pedagogical example, consider an implementation of PowerLaw:

# define the model and parameters
@with_kw struct PowerLaw{F1,F2} <: AbstractSpectralModel
    "Normalisation."
    K::F1 = FitParam(1.0)
    "Photon index."
    a::F2 = FitParam(0.5)
end

# tell the package what kind of model this is
modelkind(::Type{<:PowerLaw}) = Additive()

# define how the model acts: 
# all of the parameters are passed, in order, as arguments
function SpectralFitting.invoke!(flux, energy, ::Type{<:PowerLaw}, a)
    α = 1 - a
    α⁻¹ = inv(α)
    SpectralFitting.finite_diff_kernel!(flux, energy) do E
        α⁻¹ * E^α
    end
end

Note that each parameter has its own parametric type, here ::F1 and ::F2. This is so that the type of fit parameters can be changed, or alternative number types inserted (see Why & how). The standard parameter type is FitParam, which is what we have used here.

The model is defined to be additive by overloading modelkind.

Danger

Additive models must have a normalisation parameter with the symbol K, which is, however, not passed to SpectralFitting.invoke!. Multiplicative and convolutional models have no such requirements.

We have also used the SpectralFitting.finite_diff_kernel! utility function, which is designed to help us efficiently compute finite-difference integrations of flux, just as XSPEC does.

Note

To add additional XSPEC models, see Wrapping new XSPEC models.

XSPEC models

SpectralFitting.XS_PowerLaw — Type

XS_PowerLaw(K, a)

K

Normalisation. Default: FitParam(1.0)

a

Photon index. Default: FitParam(0.5)

Example

energy = collect(range(0.1, 20.0, 100))
invokeflux(energy, XS_PowerLaw())

                      XS_PowerLaw
       ┌────────────────────────────────────────┐
   0.5 │                                        │
       │:                                       │
       │:                                       │
       │:                                       │
       │:                                       │
       │:                                       │
       │:                                       │
       │ :                                      │
       │ :                                      │
       │  :                                     │
       │   :.                                   │
       │    ':..                                │
       │        ''':......                      │
       │                  ''''''''''''''........│
     0 │                                        │
       └────────────────────────────────────────┘
        0                                     20
                         E (keV)

SpectralFitting.XS_BlackBody — Type

XS_BlackBody(K, T)

K

Normalisation. Default: FitParam(1.0)

T

Temperature (keV). Default: FitParam(3.0)

Example

energy = collect(range(0.1, 20.0, 100))
invokeflux(energy, XS_BlackBody())

                      XS_BlackBody
       ┌────────────────────────────────────────┐
   0.2 │                                        │
       │                                        │
       │                                        │
       │                                        │
       │                                        │
       │                                        │
       │      .:''':..                          │
       │    .:       ''.                        │
       │   .'           ':.                     │
       │   :              ''..                  │
       │  :                  ':.                │
       │ :                     '':.             │
       │.:                         ''..         │
       │:                              '':....  │
     0 │'                                    '''│
       └────────────────────────────────────────┘
        0                                     20
                         E (keV)

SpectralFitting.XS_BremsStrahlung — Type

XS_BremsStrahlung(K, T)

K

Normalisation. Default: FitParam(1.0)

T

Plasma temperature (keV). Default: FitParam(7.0)

Example

energy = collect(range(0.1, 20.0, 100))
invokeflux(energy, XS_BremsStrahlung())

                  XS_BremsStrahlung
     ┌────────────────────────────────────────┐
   2 │                                        │
     │.                                       │
     │:                                       │
     │:                                       │
     │:                                       │
     │:                                       │
     │:                                       │
     │:                                       │
     │:                                       │
     │:                                       │
     │:                                       │
     │:                                       │
     │'.                                      │
     │ :                                      │
   0 │  ':....................................│
     └────────────────────────────────────────┘
      0                                     20
                       E (keV)

SpectralFitting.XS_Laor — Type

XS_Laor(K, lineE, a, inner_r, outer_r, incl)

K

Normalisation. Default: FitParam(1.0)

lineE

Rest frame line energy (keV). Default: FitParam(6.4)

a

Power law dependence of emissitivy. Scales R⁻ᵅ. Default: FrozenFitParam(3.0)

inner_r

Inner radius of the accretion disk (GM/c). Default: FrozenFitParam(1.235)

outer_r

Outer radius of the accretion disk (GM/c). Default: FrozenFitParam(400.0)

incl

Disk inclination angle to line of sight (degrees, 0 is pole on). Default: FitParam(30.0)

Example

energy = collect(range(0.1, 10.0, 100))
invokeflux(energy, XS_Laor())

                          XS_Laor
        ┌────────────────────────────────────────┐
   0.06 │                                        │
        │                                        │
        │                         ::             │
        │                         ::             │
        │                        : :             │
        │                       :  :             │
        │                      :   :             │
        │                     :'   :             │
        │                   .:     :             │
        │                  :'      :             │
        │                .'        :             │
        │              .:'         :             │
        │            ..'           :             │
        │          .:'              :            │
      0 │.......:''                 :............│
        └────────────────────────────────────────┘
         0                                     10
                          E (keV)

SpectralFitting.XS_DiskLine — Type

XS_DiskLine(K, lineE, β, inner_r, outer_r, incl)

K

Normalisation. Default: FitParam(1.0)

lineE

Rest frame line energy (keV). Default: FitParam(6.7)

β

Power law dependence of emissitivy. If < 10, scales Rᵅ. Default: FrozenFitParam(-2.0)

inner_r

Inner radius of the accretion disk (GM/c). Default: FrozenFitParam(10.0)

outer_r

Outer radius of the accretion disk (GM/c). Default: FrozenFitParam(1000.0)

incl

Disk inclination angle to line of sight (degrees, 0 is pole on). Default: FitParam(30.0)

Example

energy = collect(range(4.0, 8.0, 100))
invokeflux(energy, XS_DiskLine())

                        XS_DiskLine
        ┌────────────────────────────────────────┐
   0.09 │                                        │
        │                           .            │
        │                           :            │
        │                           ::           │
        │                         . ::           │
        │                         : ::           │
        │                         :'':           │
        │                        .'  :           │
        │                        :    :          │
        │                        :    :          │
        │                       .'    :          │
        │                       :     :          │
        │                     .:      '.         │
        │                   .:'        :         │
      0 │...............:'''           :.........│
        └────────────────────────────────────────┘
         4                                      8
                          E (keV)

SpectralFitting.XS_PhotoelectricAbsorption — Type

XS_PhotoelectricAbsorption(ηH)

ηH

Equivalent hydrogen column (units of 10²² atoms per cm⁻²). Default: FitParam(1.0)

Example

energy = collect(range(0.1, 20.0, 100))
invokeflux(energy, XS_PhotoelectricAbsorption())

             XS_PhotoelectricAbsorption
     ┌────────────────────────────────────────┐
   1 │       ...''''''''''''''''''''''''''''''│
     │      .'                                │
     │     :                                  │
     │    :'                                  │
     │    :                                   │
     │   :                                    │
     │   :                                    │
     │   :                                    │
     │  :                                     │
     │  :                                     │
     │  :                                     │
     │  :                                     │
     │  :                                     │
     │ :                                      │
   0 │.:                                      │
     └────────────────────────────────────────┘
      0                                     20
                       E (keV)

SpectralFitting.XS_WarmAbsorption — Type

XS_WarmAbsorption(ηH, Ew)

ηH

Equivalent hydrogen column (units of 10²² atoms per cm⁻²). Default: FitParam(1.0)

Ew

Window energy (keV). Default: FitParam(1.0)

Example

energy = collect(range(0.1, 20.0, 100))
invokeflux(energy, XS_WarmAbsorption())

                    XS_WarmAbsorption
       ┌────────────────────────────────────────┐
     1 │':      ...''':'''''''''''''''''''''''''│
       │ :    .:'                               │
       │ :   .'                                 │
       │ :  .:                                  │
       │ :  :                                   │
       │ :  :                                   │
       │ : :                                    │
       │ : :                                    │
       │ : :                                    │
       │ : :                                    │
       │ ::                                     │
       │ ::                                     │
       │  :                                     │
       │  :                                     │
   0.2 │  :                                     │
       └────────────────────────────────────────┘
        0                                     20
                         E (keV)

SpectralFitting.XS_CalculateFlux — Type

XS_CalculateFlux(E_min, E_max, lg10Flux)

E_min

Minimum energy. Default: FrozenFitParam(0.2)

E_max

Maximum energy. Default: FrozenFitParam(2.0)

lg10Flux

log (base 10) flux in erg / cm^2 / s Default: FitParam(-10.0, lowerlimit = -Inf, upperlimit = 0.0)

Wrapping new XSPEC models

SpectralFitting exports a helpful macro to facilitate wrapping additional XSPEC models.

SpectralFitting.@xspecmodel — Macro

@xspecmodel model_kind func_name model

Used to wrap additional XSPEC models, generating the needed AbstractSpectralModel implementation.

Examples

@xspecmodel Additive :C_powerlaw struct XS_PowerLaw{F1,F2}
    "Normalisation."
    K::F1 = FitParam(1.0)
    "Photon index."
    a::F2 = FitParam(0.5)
end

We define a new structure XS_PowerLaw with two parameters, but since the model is Additive, only a single parameter (a) is passed to the XSPEC function. The function we bind to this model is :C_powerlaw from the XSPEC C wrappers.

The macro will then generate the following functions

modelkind(::Type{<:XS_PowerLaw})
implementation(::Type{<:XS_PowerLaw})
invoke!(::Type{<:XS_PowerLaw})

Generating model fingerprints

To generate the unicode plot to add as a fingerprint, we use a simple function:

using UnicodePlots

function plotmodel(energy, model)
    flux = invokemodel(energy, model)
    lineplot(
        energy[1:end-1],
        flux,
        title=String(SpectralFitting.model_base_name(typeof(model))),
        xlabel="E (keV)",
        canvas=DotCanvas
    )
end

# e.g. for XS_PowerLaw()
energy = collect(range(0.1, 20.0, 100))
plotmodel(energy, XS_PowerLaw())

                       XS_PowerLaw                
       ┌────────────────────────────────────────┐ 
   0.5 │                                        │ 
       │:                                       │ 
       │:                                       │ 
       │:                                       │ 
       │:                                       │ 
       │:                                       │ 
       │:                                       │ 
       │ :                                      │ 
       │ :                                      │ 
       │  :                                     │ 
       │   :.                                   │ 
       │    ':..                                │ 
       │        ''':......                      │ 
       │                  ''''''''''''''........│ 
     0 │                                        │ 
       └────────────────────────────────────────┘ 
        0                                     20  
                         E (keV)