Models index

Models wrapped from XSPEC implementations are prefixed with XS_*, whereas pure-Julia models are simply named, e.g. XS_PowerLaw in XSPEC vs PowerLaw in Julia.

The available models are

SpectralFitting.BlackBody

SpectralFitting.PowerLaw

SpectralFitting.XS_BlackBody

SpectralFitting.XS_BremsStrahlung

SpectralFitting.XS_CalculateFlux

SpectralFitting.XS_DiskLine

SpectralFitting.XS_Gaussian

SpectralFitting.XS_KerrDisk

SpectralFitting.XS_KyrLine

SpectralFitting.XS_Laor

SpectralFitting.XS_PhotoelectricAbsorption

SpectralFitting.XS_PowerLaw

SpectralFitting.XS_WarmAbsorption

Julia models

SpectralFitting.PowerLaw — Type

XS_PowerLaw(K, a)

K: Normalisation.

a: Photon index.

Example

energy = collect(range(0.1, 20.0, 100))
invokemodel(energy, PowerLaw())

                        PowerLaw
       ┌────────────────────────────────────────┐
   0.5 │                                        │
       │:                                       │
       │:                                       │
       │:                                       │
       │:                                       │
       │:                                       │
       │:                                       │
       │ :                                      │
       │ :                                      │
       │  :                                     │
       │   :.                                   │
       │    ':..                                │
       │        ''':......                      │
       │                  ''''''''''''''........│
     0 │                                        │
       └────────────────────────────────────────┘
        0                                     20
                         E (keV)

SpectralFitting.BlackBody — Type

XS_BlackBody(K, T)

K: Normalisation.

kT: Temperature (keV).

Example

energy = collect(range(0.1, 20.0, 100))
invokemodel(energy, BlackBody())

                        BlackBody
       ┌────────────────────────────────────────┐
   0.2 │                                        │
       │                                        │
       │                                        │
       │                                        │
       │                                        │
       │                                        │
       │      .:''':..                          │
       │     :'      '':.                       │
       │   .'           ':.                     │
       │  .:               '..                  │
       │  :                  ':.                │
       │ .'                     ':..            │
       │ :                         ''...        │
       │:                              '''....  │
     0 │:                                    '''│
       └────────────────────────────────────────┘
        0                                     20
                         E (keV)

Defining new models

To define a new Julia model, we need only implement the AbstractSpectralModel interface. As a pedagogical example, consider an implementation of PowerLaw:

# define the model and parameters
@with_kw struct PowerLaw{F1,F2} <: AbstractSpectralModel
    "Normalisation."
    K::F1 = FitParam(1.0)
    "Photon index."
    a::F2 = FitParam(0.5)
end

# tell the package what kind of model this is
modelkind(::Type{<:PowerLaw}) = Additive()

# define how the model acts: 
# all of the parameters are passed, in order, as arguments
function SpectralFitting.invoke!(flux, energy, ::Type{<:PowerLaw}, a)
    α = 1 - a
    α⁻¹ = inv(α)
    SpectralFitting.finite_diff_kernel!(flux, energy) do E
        α⁻¹ * E^α
    end
end

Note that each parameter has its own parametric type, here ::F1 and ::F2. This is so that the type of fit parameters can be changed, or alternative number types inserted (see Why & how). The standard parameter type is FitParam, which is what we have used here.

The model is defined to be additive by overloading modelkind.

Danger

Additive models must have a normalisation parameter with the symbol K, which is, however, not passed to SpectralFitting.invoke!. Multiplicative and convolutional models have no such requirements.

We have also used the SpectralFitting.finite_diff_kernel! utility function, which is designed to help us efficiently compute finite-difference integrations of flux, just as XSPEC does.

Note

To add additional XSPEC models, see Wrapping new XSPEC models.

XSPEC models

XSPEC models frequently have tabular data dependencies, without which the models fail to invoke (see Model data availability). If the data files are known but not present, the XSPEC models will throw an error with instructions for downloading the missing data. If the data files are unknown, Julia may crash catastrophically. If this is the case, often a single line will be printed with the LibXSPEC error, specifying the name of the missing source file. This can be registered as a data dependency of a model using SpectralFitting.register_model_data.

The first time any XSPEC model is invoked, SpectralFitting checks to see whether requisite data is needed, and whether the data is downloaded. Subsequent calls will hit a lookup cache instead to avoid run-time costs of performing this check.

SpectralFitting.XS_PowerLaw — Type

XS_PowerLaw(K, a)

K: Normalisation.

a: Photon index.

Example

energy = collect(range(0.1, 20.0, 100))
invokemodel(energy, XS_PowerLaw())

                      XS_PowerLaw
       ┌────────────────────────────────────────┐
   0.5 │                                        │
       │:                                       │
       │:                                       │
       │:                                       │
       │:                                       │
       │:                                       │
       │:                                       │
       │ :                                      │
       │ :                                      │
       │  :                                     │
       │   :.                                   │
       │    ':..                                │
       │        ''':......                      │
       │                  ''''''''''''''........│
     0 │                                        │
       └────────────────────────────────────────┘
        0                                     20
                         E (keV)

SpectralFitting.XS_BlackBody — Type

XS_BlackBody(K, T)

K: Normalisation.

T: Temperature (keV).

Example

energy = collect(range(0.1, 20.0, 100))
invokemodel(energy, XS_BlackBody())

                      XS_BlackBody
       ┌────────────────────────────────────────┐
   0.2 │                                        │
       │                                        │
       │                                        │
       │                                        │
       │                                        │
       │                                        │
       │      .:''':..                          │
       │    .:       ''.                        │
       │   .'           ':.                     │
       │   :              ''..                  │
       │  :                  ':.                │
       │ :                     '':.             │
       │.:                         ''..         │
       │:                              '':....  │
     0 │'                                    '''│
       └────────────────────────────────────────┘
        0                                     20
                         E (keV)

SpectralFitting.XS_BremsStrahlung — Type

XS_BremsStrahlung(K, T)

K: Normalisation.

T: Plasma temperature (keV).

Example

energy = collect(range(0.1, 20.0, 100))
invokemodel(energy, XS_BremsStrahlung())

                  XS_BremsStrahlung
     ┌────────────────────────────────────────┐
   2 │                                        │
     │.                                       │
     │:                                       │
     │:                                       │
     │:                                       │
     │:                                       │
     │:                                       │
     │:                                       │
     │:                                       │
     │:                                       │
     │:                                       │
     │:                                       │
     │'.                                      │
     │ :                                      │
   0 │  ':....................................│
     └────────────────────────────────────────┘
      0                                     20
                       E (keV)

SpectralFitting.XS_Gaussian — Type

XS_Gaussian(K, E, σ)

K: Normalisation

E: Line wavelength in Angstrom.

σ: Line width in Angstrom.

Example

energy = collect(range(4.0, 8.0, 100))
invokemodel(energy, XS_Gaussian())

                        XS_Gaussian                
        ┌────────────────────────────────────────┐ 
   0.09 │                                        │ 
        │            .                           │ 
        │           : :                          │ 
        │           : :                          │ 
        │           : '.                         │ 
        │          .'  :                         │ 
        │          :   :                         │ 
        │          :   :                         │ 
        │          :   '.                        │ 
        │         :     :                        │ 
        │         :     :                        │ 
        │         :     :                        │ 
        │        .'      :                       │ 
        │        :       :                       │ 
      0 │.......:         :......................│ 
        └────────────────────────────────────────┘ 
         0                                     20  
                          E (keV)                  

SpectralFitting.XS_Laor — Type

XS_Laor(K, lineE, a, inner_r, outer_r, incl)

K: Normalisation.

lineE: Rest frame line energy (keV).

a: Power law dependence of emissitivy. Scales R⁻ᵅ.

inner_r: Inner radius of the accretion disk (GM/c).

outer_r: Outer radius of the accretion disk (GM/c).

θ: Disk inclination angle to line of sight (degrees, 0 is pole on).

Example

energy = collect(range(0.1, 10.0, 100))
invokemodel(energy, XS_Laor())

                          XS_Laor
        ┌────────────────────────────────────────┐
   0.06 │                                        │
        │                                        │
        │                         ::             │
        │                         ::             │
        │                        : :             │
        │                       :  :             │
        │                      :   :             │
        │                     :'   :             │
        │                   .:     :             │
        │                  :'      :             │
        │                .'        :             │
        │              .:'         :             │
        │            ..'           :             │
        │          .:'              :            │
      0 │.......:''                 :............│
        └────────────────────────────────────────┘
         0                                     10
                          E (keV)

SpectralFitting.XS_DiskLine — Type

XS_DiskLine(K, lineE, β, inner_r, outer_r, incl)

K: Normalisation.

lineE: Rest frame line energy (keV).

β: Power law dependence of emissitivy. If < 10, scales Rᵅ.

inner_r: Inner radius of the accretion disk (GM/c).

outer_r: Outer radius of the accretion disk (GM/c).

θ: Disk inclination angle to line of sight (degrees, 0 is pole on).

Example

energy = collect(range(4.0, 8.0, 100))
invokemodel(energy, XS_DiskLine())

                        XS_DiskLine
        ┌────────────────────────────────────────┐
   0.09 │                                        │
        │                           .            │
        │                           :            │
        │                           ::           │
        │                         . ::           │
        │                         : ::           │
        │                         :'':           │
        │                        .'  :           │
        │                        :    :          │
        │                        :    :          │
        │                       .'    :          │
        │                       :     :          │
        │                     .:      '.         │
        │                   .:'        :         │
      0 │...............:'''           :.........│
        └────────────────────────────────────────┘
         4                                      8
                          E (keV)

SpectralFitting.XS_PhotoelectricAbsorption — Type

XS_PhotoelectricAbsorption(ηH)

ηH: Equivalent hydrogen column (units of 10²² atoms per cm⁻²).

Example

energy = collect(range(0.1, 20.0, 100))
invokemodel(energy, XS_PhotoelectricAbsorption())

             XS_PhotoelectricAbsorption
     ┌────────────────────────────────────────┐
   1 │       ...''''''''''''''''''''''''''''''│
     │      .'                                │
     │     :                                  │
     │    :'                                  │
     │    :                                   │
     │   :                                    │
     │   :                                    │
     │   :                                    │
     │  :                                     │
     │  :                                     │
     │  :                                     │
     │  :                                     │
     │  :                                     │
     │ :                                      │
   0 │.:                                      │
     └────────────────────────────────────────┘
      0                                     20
                       E (keV)

SpectralFitting.XS_WarmAbsorption — Type

XS_WarmAbsorption(ηH, Ew)

ηH: Equivalent hydrogen column (units of 10²² atoms per cm⁻²).

Ew: Window energy (keV).

Example

energy = collect(range(0.1, 20.0, 100))
invokemodel(energy, XS_WarmAbsorption())

                    XS_WarmAbsorption
       ┌────────────────────────────────────────┐
     1 │':      ...''':'''''''''''''''''''''''''│
       │ :    .:'                               │
       │ :   .'                                 │
       │ :  .:                                  │
       │ :  :                                   │
       │ :  :                                   │
       │ : :                                    │
       │ : :                                    │
       │ : :                                    │
       │ : :                                    │
       │ ::                                     │
       │ ::                                     │
       │  :                                     │
       │  :                                     │
   0.2 │  :                                     │
       └────────────────────────────────────────┘
        0                                     20
                         E (keV)

SpectralFitting.XS_CalculateFlux — Type

XS_CalculateFlux(E_min, E_max, lg10Flux)

E_min: Minimum energy.

E_max: Maximum energy.

log10Flux: log (base 10) flux in erg / cm^2 / s

SpectralFitting.XS_KerrDisk — Type

XS_KerrDisk(K, lineE, index1, index2, break_r, a, θ, inner_r, outer_r)

K: Normalisation.

lineE: Rest frame line energy (keV).

index1: Emissivity index for inner disk.

index2: Emissivity index for outer disk.

break_r: Break radius seperating inner and outer disk (gᵣ).

a: Dimensionless black hole spin.

θ: Disk inclination angle to line of sight (degrees).

inner_r: Inner radius of the disk in units of rₘₛ.

outer_r: Outer radius of the disk in units of rₘₛ.

z: Redshift.

Example

energy = collect(range(0.1, 20.0, 100))
invokemodel(energy, XS_KerrDisk())

                        XS_KerrDisk
        ┌────────────────────────────────────────┐
   0.05 │                                        │
        │                                        │
        │                                .       │
        │                               .:       │
        │                              :::       │
        │                            .:' '.      │
        │                           .:    :      │
        │                         ..'     :      │
        │                         :'      :      │
        │                       .'        :      │
        │                     .:'         :      │
        │                  .:''           :      │
        │               .::'              :      │
        │            ..:'                 :      │
      0 │.........:'''                    :......│
        └────────────────────────────────────────┘
         0                                      8
                          E (keV)

SpectralFitting.XS_KyrLine — Type

XS_KyrLine(K, a, θ_obs, inner_r, ms_flag, outer_r, lineE, α, β, break_r, z, limb)

K: Normalisation.

a: Dimensionless black hole spin.

θ: Observer inclination (0 is on pole, degrees).

inner_r: Inner radius of the disk in units of GM/c²

ms_flag: 0: integrate from rᵢₙ. 1: integrate from rₘₛ.

outer_r: Outer radius of the disk in units of GM/c²

lineE: Rest frame line energy (keV).

α

β

break_r: Break radius seperating inner and outer disk (GM/c²).

z: Overall Doppler shift.

limb: 0: isotropic emission, 1: Laor's limb darkening, 2: Haard's limb brightening.

Example

energy = collect(range(0.1, 20.0, 100))
invokemodel(energy, XS_KyrLine())

                        XS_KyrLine
        ┌────────────────────────────────────────┐
   0.05 │                                        │
        │                                        │
        │                                :       │
        │                                :.      │
        │                              :.':      │
        │                             :'  :      │
        │                            :    :      │
        │                          .'     :      │
        │                        .:'      :      │
        │                       .'        :      │
        │                     .:          :      │
        │                   .'            :      │
        │                .:'              :      │
        │            ..:'                 :      │
      0 │.........:'''                    :......│
        └────────────────────────────────────────┘
         0                                      8
                          E (keV)

Wrapping new XSPEC models

SpectralFitting exports a helpful macro to facilitate wrapping additional XSPEC models.

SpectralFitting.@xspecmodel — Macro

@xspecmodel model_kind func_name model

Used to wrap additional XSPEC models, generating the needed AbstractSpectralModel implementation.

Examples

@xspecmodel Additive :C_powerlaw struct XS_PowerLaw{F1,F2}
    "Normalisation."
    K::F1 = FitParam(1.0)
    "Photon index."
    a::F2 = FitParam(0.5)
end

We define a new structure XS_PowerLaw with two parameters, but since the model is Additive, only a single parameter (a) is passed to the XSPEC function. The function we bind to this model is :C_powerlaw from the XSPEC C wrappers.

The macro will then generate the following functions

modelkind(::Type{<:XS_PowerLaw})
implementation(::Type{<:XS_PowerLaw})
invoke!(::Type{<:XS_PowerLaw})

SpectralFitting.register_model_data — Function

SpectralFitting.register_model_data(M::Type{<:AbstractSpectralModel}, filenames::String...)
SpectralFitting.register_model_data(s::Symbol, filenames::String...)

Register filenames as model data associated with the model given by type M or symbol s. This function does not download any files, but rather adds the relevant filenames to a lookup which SpectralFitting.download_model_data consults when invoked, and consequently model data is only downloaded when needed.

Note

It is good practice to use this method immediately after defining a new model with @xspecmodel to register any required datafiles from the HEASoft source code, and therefore keep relevant information together.

Example

# by type
register_model_data(XS_Laor, "ari.mod")
# by symbol
register_model_data(:XS_KyrLine, "KBHline01.fits")

Generating model fingerprints

To generate the unicode plot to add as a fingerprint, we use a simple function:

using UnicodePlots

function plotmodel(energy, model)
    flux = invokemodel(energy, model)
    lineplot(
        energy[1:end-1], 
        flux, 
        title=String(SpectralFitting.model_base_name(typeof(model))), 
        xlabel="E (keV)", 
        canvas=DotCanvas
    )
end

# e.g. for XS_PowerLaw()
energy = collect(range(0.1, 20.0, 100))
plotmodel(energy, XS_PowerLaw())