Model index

Models wrapped from XSPEC implementations are prefixed with XS_*, whereas pure-Julia models are simply named, e.g. XS_PowerLaw in XSPEC vs PowerLaw in Julia.

The available models are

SpectralFitting.BlackBody

SpectralFitting.BremsStrahlung

SpectralFitting.DeltaLine

SpectralFitting.GaussianLine

SpectralFitting.Log10Flux

SpectralFitting.PhotoelectricAbsorption

SpectralFitting.PowerLaw

Julia models

Additive

SpectralFitting.BlackBody — Type

BlackBody

K: Normalisation.

kT: Temperature (keV).

Example

energy = collect(range(0.1, 20.0, 100))
invokemodel(energy, BlackBody())

                        BlackBody
       ┌────────────────────────────────────────┐
   0.2 │                                        │
       │                                        │
       │                                        │
       │                                        │
       │                                        │
       │                                        │
       │      .:''':..                          │
       │     :'      '':.                       │
       │   .'           ':.                     │
       │  .:               '..                  │
       │  :                  ':.                │
       │ .'                     ':..            │
       │ :                         ''...        │
       │:                              '''....  │
     0 │:                                    '''│
       └────────────────────────────────────────┘
        0                                     20
                         E (keV)

SpectralFitting.BremsStrahlung — Type

BremsStrahlung

K: Normalisation.

kT: Temperature (keV).

ab: Helium to hydrogen ratio

Example

energy = collect(range(0.1, 20.0, 100))
invokemodel(energy, BremsStrahlung())

                   BremsStrahlung
     ┌────────────────────────────────────────┐
   3 │.                                       │
     │:                                       │
     │:                                       │
     │:                                       │
     │:                                       │
     │:                                       │
     │:                                       │
     │:                                       │
     │:                                       │
     │:                                       │
     │:                                       │
     │:                                       │
     │:                                       │
     │ :                                      │
   0 │ ':.....................................│
     └────────────────────────────────────────┘
      0                                     20
                       E (keV)

SpectralFitting.DeltaLine — Type

DeltaLine

_width

K: Normalisation.

E: Energy at which the delta function spikes.

Example

energy = collect(range(0.1, 20.0, 100))
invokemodel(energy, DeltaLine())

                        DeltaLine
       ┌────────────────────────────────────────┐
   0.4 │         .                              │
       │         :                              │
       │         :                              │
       │         :                              │
       │         :                              │
       │         :                              │
       │         ::                             │
       │         ::                             │
       │         ::                             │
       │         ::                             │
       │         ::                             │
       │         ::                             │
       │         ::                             │
       │         ::                             │
     0 │.........::.............................│
       └────────────────────────────────────────┘
        0                                     20
                         E (keV)

Note

The DeltaLine model is not a true delta function, as this would be extremely difficult to define in a numerical model that needs to be able to propagate gradients. Instead, it is a very narrow GaussianLine model.

SpectralFitting.GaussianLine — Type

GaussianLine

K: Normalisation.

μ: Mean

σ: Standard deviation

Example

energy = collect(range(0.1, 20.0, 100))
invokemodel(energy, GuassianLine())

                       GaussianLine
        ┌────────────────────────────────────────┐
   0.09 │                                        │
        │            ..                          │
        │           .':                          │
        │           : :                          │
        │           : '.                         │
        │          .'  :                         │
        │          :   :                         │
        │          :   :                         │
        │          :   '.                        │
        │         :     :                        │
        │         :     :                        │
        │         :     :                        │
        │        .'      :                       │
        │        :       '.                      │
      0 │.......:         :......................│
        └────────────────────────────────────────┘
         0                                     20
                          E (keV)

SpectralFitting.PowerLaw — Type

PowerLaw

K: Normalisation.

a: Photon index.

Example

energy = collect(range(0.1, 20.0, 100))
invokemodel(energy, PowerLaw())

                        PowerLaw
       ┌────────────────────────────────────────┐
   0.5 │                                        │
       │:                                       │
       │:                                       │
       │:                                       │
       │:                                       │
       │:                                       │
       │:                                       │
       │ :                                      │
       │ :                                      │
       │  :                                     │
       │   :.                                   │
       │    ':..                                │
       │        ''':......                      │
       │                  ''''''''''''''........│
     0 │                                        │
       └────────────────────────────────────────┘
        0                                     20
                         E (keV)

Multiplicative

SpectralFitting.PhotoelectricAbsorption — Type

PhotoelectricAbsorption

table

ηH: Equivalent hydrogen column (units of 10²² atoms per cm⁻²).

Convolutional

SpectralFitting.Log10Flux — Type

Log10Flux

Used to measure the (log) flux of the models it is applied to. Note that the additive components must have their normalisations frozen for this model to work properly.

E_min

E_max

log10Flux

Example

model = PowerLaw()
model.K.frozen = true

flux_model = Log10Flux()(model)

Utility

SpectralFitting.register_model_data — Function

SpectralFitting.register_model_data(M::Type{<:AbstractSpectralModel}, model_data::ModelDataInfo...)
SpectralFitting.register_model_data(M::Type{<:AbstractSpectralModel}, remote_and_local::Tuple{String,String}...)
SpectralFitting.register_model_data(M::Type{<:AbstractSpectralModel}, filenames::String...)
SpectralFitting.register_model_data(s::Symbol, filenames::String...)

Register filenames as model data associated with the model given by type M or symbol s. This function does not download any files, but rather adds the relevant filenames to a lookup which SpectralFitting.download_model_data consults when invoked, and consequently model data is only downloaded when needed.

Note

It is good practice to use this method immediately after defining a new model with @xspecmodel to register any required datafiles from the HEASoft source code, and therefore keep relevant information together.

Example

# by type
register_model_data(XS_Laor, "ari.mod")
# by symbol
register_model_data(:XS_KyrLine, "KBHline01.fits")

SpectralFitting.finite_diff_kernel! — Function

finite_diff_kernel!(f::Function, flux, energy)

Calculates the finite difference of the function f over the energy bin between the high and low bin edges, via

\[c_i = f(E_{i,\text{high}}) - f(E_{i,\text{low}}),\]

similar to evaluating the limits of the integral between $E_{i,\text{high}}$ and $E_{i,\text{low}}$.

This utility function is primarily used for Additive models to ensure the flux per bin is normalised for the energy over the bin.

SpectralFitting.wrap_model_as_objective — Function

wrap_model_as_objective(model::AbstractSpectralModel; ΔE = 1e-1)
wrap_model_as_objective(M::Type{<:AbstractSpectralModel}; ΔE = 1e-1)

Wrap a spectral model into an objective function for building/optimizing a surrogate model. Returns an anonymous function taking the tuple (E, params...) as the argument, and returning a single flux value.

Generating model fingerprints

To generate the unicode plot to add as a fingerprint, we use a simple function:

using SpectralFitting, UnicodePlots

function plotmodel(energy, model)
    flux = invokemodel(energy, model)
    lineplot(
        energy[1:end-1],
        flux,
        title=String(Base.typename(typeof(model)).name),
        xlabel="E (keV)",
        canvas=DotCanvas
    )
end

# e.g. for XS_PowerLaw()
energy = collect(range(0.1, 20.0, 100))
plotmodel(energy, PowerLaw())

                      PowerLaw                  
     ┌────────────────────────────────────────┐ 
   7 │.                                       │ 
     │:                                       │ 
     │:                                       │ 
     │:                                       │ 
     │:                                       │ 
     │:                                       │ 
     │:                                       │ 
     │:                                       │ 
     │:                                       │ 
     │:                                       │ 
     │:                                       │ 
     │:                                       │ 
     │:                                       │ 
     │:.                                      │ 
   0 │ :......................................│ 
     └────────────────────────────────────────┘ 
      0                                     20  
                       E (keV)