API reference

Utility functions for defining new models

SpectralFitting.finite_diff_kernel! — Function

finite_diff_kernel!(f::Function, flux, energy)

Calculates the finite difference of the function f over the energy bin between the high and low bin edges, via

\[c_i = f(E_{i,\text{high}}) - f(E_{i,\text{low}}),\]

similar to evaluating the limits of the integral between $E_{i,\text{high}}$ and $E_{i,\text{low}}$.

This utility function is primarily used for Additive models to ensure the flux per bin is normalised for the energy over the bin.

SpectralFitting.wrap_model_as_objective — Function

wrap_model_as_objective(model::AbstractSpectralModel; ΔE = 1e-1)
wrap_model_as_objective(M::Type{<:AbstractSpectralModel}; ΔE = 1e-1)

Wrap a spectral model into an objective function for building/optimizing a surrogate model. Returns an anonymous function taking the tuple (E, params...) as the argument, and returning a single flux value.

General reference

Base.copy — Method

Base.copy(m::AbstractTableModel)

Create a copy of an AbstractTableModel. This will copy all fields except the table field, which is assumed to be a constant set of values that can be shared by multiple copies.

When this is not the case, the user should redefine Base.copy for their particular table model to copy the table as needed.

Base.copy — Method

Base.copy(m::AsConvolution)

Creates a copy of an AsConvolution wrapped model. Will make a deepcopy of the cache to elimiate possible thread contention, but does not copy the domain.

SpectralFitting._accumulated_indices — Method

_accumulated_indices(items)

items is a tuple or vector of lengths n1, n2, ...

Returns a tuple or array with same length as items, which gives the index boundaries of an array with size n1 + n2 + ....

SpectralFitting._bind_pairs! — Method

Bind the symbols of last(pair) in all models indexes by first(pair).

SpectralFitting._convolve_irregular_grid! — Method

Assumes A is binned in X1 and kernel is binned in X2. Output will also be binned on X1

SpectralFitting._safe_ffi_invoke! — Method

function _safe_ffi_invoke!(output, input, params, ModelType::Type{<:AbstractSpectralModel})

Wrapper to do a foreign function call to an XSPEC model.

SpectralFitting._subtract_background! — Method

Does the background subtraction and returns units of counts. That means we have multiplied through by a factor $t_D$ relative to the reference equation (2.3) in the XSPEC manual.

SpectralFitting._unsafe_ffi_invoke! — Method

function _unsafe_ffi_invoke!(
    output,
    error_vec,
    input,
    params,
    ModelType::Type{<:AbstractSpectralModel},
)

Wrapper to do a foreign function call to an XSPEC model.

See also _safe_ffi_invoke!.

SpectralFitting.adjust_free_bindings — Method

adjust_free_bindings(model::FittableMultiModel, bindings::Vector{Vector{Pair{Int,Int}}})

Returns a new parameter binding list with the parameter indices adjusted to omit the frozen parameter. That is, if a model has three parameters a, b (frozen) and c, and parameter c (index 3) is bound; then the new bindings will instead give the index of c as 2.

In that sense, the new bindings refer to the free parameter vector, and not the full parameter vector.

If the binding refers to a frozen parameter, it is removed from the binding list.

SpectralFitting.allocate_model_output — Method

allocate_model_output(model::AbstractSpectralModel, domain::AbstractVector)

Allocate the output space for the AbstractSpectralModel for a given domain. The output type will be the promoted element type of the model and domain.

Uses construct_objective_cache to construct the appropriate layout.

SpectralFitting.bind! — Method

bind!(prob::FittingProblem, pairs::Pair{Int,Symbol}...)
bind!(prob::FittingProblem, symbols::Symbol...)

Bind parameters together within a FittingProblem. Parameters bound together will be mandated to have exact same value during the fit.

The binding may either be a single symbol that is present in all models in the fitting problem, or a series of pairs Int => Symbol which index the specific model and parameters to bind together. All bindings specified in a single call to bind! will be bound together. Multiple bindings are possible with repeated call to bind!.

Bind model 1's K_1 parameter to model 2's K_3:

bind!(prob, 1 => :K_1, 2 => :K_3)

Bind model 3's K_2 parameter to model4's :L_1 and model 6's a_3:

bind!(prob, 3 => :K_2, 4 => :L_1, 6 => :a_3)

Bind the K_1 parameter across all the models:

bind!(prob, :K_1)

Examples

Consider the following two models

model1 = PhotoelectricAbsorption() * (BlackBody() + PowerLaw())
model2 = PhotoelectricAbsorption() * (PowerLaw() + PowerLaw())

prob = FittingProblem(model1 => data1, model2 => data2)

# bind the power law indices in the two models
bind!(prob, :a_1)

# bind the normalisation of powerlaws in the 2nd model:
bind!(prob, 2 => :K_1, 2 => :K_2)

Note

Only free parameters can be bound together.

SpectralFitting.common_support — Method

common_support(x, y)

Find the common AbstractLayout of x and y, following the ordering of preferred_support.

SpectralFitting.count_error — Method

count_error(k, σ)

Gives the error on k mean photon counts to a significance of σ standard deviations about the mean.

Derived from likelihood of binomial distributions being the beta function.

SpectralFitting.details — Method

details(prob::FittingProblem)

Show details about the fitting problem, including the specific model parameters that are bound together.

SpectralFitting.download_all_model_data — Method

SpectralFitting.download_all_model_data()

Downloads all model data for the models currently registered with SpectralFitting.register_model_data. Calls SpectralFitting.download_model_data to perform the download.

SpectralFitting.download_model_data — Method

SpectralFitting.download_model_data(model::AbstractSpectralModel; kwargs...)
SpectralFitting.download_model_data(M::Type{<:AbstractSpectralModel}; kwargs...)
SpectralFitting.download_model_data(s::Symbol; kwargs...)

Downloads the model data for a model specified either by model, type M, or symbol s. Datafiles associated with a specific model may be registered using SpectralFitting.register_model_data. The download is currently unconfigurable, but permits slight control via a number of keyword arguments:

progress::Bool = true

Display a progress bar for the download.

model_source_url::String = "http://www.star.bris.ac.uk/fbaker/XSPEC-model-data"

The source URL used to download the model data.

All standard XSPEC spectral model data is currently being hosted on the University of Bristol astrophysics servers, and should be persistently available to anyone.

SpectralFitting.error_statistic — Method

error_statistic(::AbstractDataset)

Should return an ErrorStatistics describing which error statistic this data uses.

If undefined for a derived type, returns ErrorStatistics.Unknown.

SpectralFitting.finite_diff_kernel! — Method

finite_diff_kernel!(f::Function, flux, energy)

Calculates the finite difference of the function f over the energy bin between the high and low bin edges, via

\[c_i = f(E_{i,\text{high}}) - f(E_{i,\text{low}}),\]

similar to evaluating the limits of the integral between $E_{i,\text{high}}$ and $E_{i,\text{low}}$.

This utility function is primarily used for Additive models to ensure the flux per bin is normalised for the energy over the bin.

SpectralFitting.folded_energy — Method

folded_energy(response::ResponseMatrix)

Get the contiguously binned energy corresponding to the output (folded) domain of the response matrix. That is, the channel energies as used by the spectrum.

SpectralFitting.get_tickslogscale — Method

get_tickslogscale(lims; skiplog=false)

Return a tuple (ticks, ticklabels) for the axis limit lims where multiples of 10 are major ticks with label and minor ticks have no label skiplog argument should be set to true if lims is already in log scale.

Modified from https://github.com/JuliaPlots/Plots.jl/issues/3318.

SpectralFitting.implementation — Method

implementation(model::AbstractSpectralModel)
implementation(::Type{<:AbstractSpectralModel})

Get the AbstractSpectralModelImplementation for a given AbstractSpectralModel or model type.

This is used primarily to learn what optimizations we can do with a model, for example propagating auto-diff gradients through a model or arbitrary precision numbers.

SpectralFitting.invoke! — Method

SpectralFitting.invoke!(output, domain, M::Type{<:AbstractSpectralModel}, params...)

Used to define the behaviour of models. Should calculate the output of the model and write in-place into output. The model parameters are passed in the model structure.

Warning

This function should not be called directly. Use invokemodel instead. invoke! is only to define the model, not to use it. Users should always call models using invokemodel or invokemodel! to ensure normalisations and closures are accounted for.

Example

Base.@kwdef struct MyModel{T} <: AbstractSpectralModel{T,Multiplicative}
    p1::T = FitParam(1.0)
    p2::T = FitParam(2.0)
    p3::T = FitParam(3.0)
end

would have the arguments passed to invoke! as

function SpectralFitting.invoke!(output, domain, model::MyModel)
    # ...
end

SpectralFitting.invokemodel! — Method

invokemodel!(output, domain, model)
invokemodel!(output, domain, model, params::AbstractVector)
invokemodel!(output, domain, model, params::ParameterCache)

In-place variant of invokemodel, calculating the output of an AbstractSpectralModel given by model, optionally overriding the parameters using a ParameterCache or an AbstractVector.

The output may not necessarily be a single vector, and one should use allocate_model_output to allocate the output structure.

Example

model = PowerLaw()
domain = collect(range(0.1, 20.0, 100))
output = allocate_model_output(model, domain)
invokemodel!(output, domain, model)

SpectralFitting.invokemodel — Method

invokemodel(domain, model::AbstractSpectralModel)

Invoke the AbstractSpectralModel given by model over the domain domain.

This function will perform any normalisation or post-processing tasks that a specific model kind may require, e.g. multiplying by a normalisation constant for Additive models.

Note

invokemodel allocates the needed output arrays based on the element type of free_params to allow automatic differentation libraries to calculate parameter gradients.

In-place non-allocating variants are the invokemodel! functions.

Example

model = PowerLaw()
domain = collect(range(0.1, 20.0, 100))

invokemodel(domain, model)

SpectralFitting.make_domain_variance — Method

make_domain_variance(layout::AbstractLayout, dataset::AbstractDataset)

Make the variance vector associated with the domain.

SpectralFitting.make_label — Method

make_label(d::AbstractDataset)

Return a string that gives a descriptive label for this dataset.

SpectralFitting.make_model_domain — Method

make_model_domain(layout::AbstractLayout, dataset::AbstractDataset)

Returns the array used as the domain for the modelling. This is paired with make_domain_variance

SpectralFitting.make_objective — Method

make_objective(layout::AbstractLayout, dataset::AbstractDataset)

Returns the array used as the target for model fitting. The array must correspond to the data AbstractLayout specified by the layout parameter.

In as far as it can be guarunteed, the memory in the returned array will not be mutated by any fitting procedures.

Domain for this objective should be returned by make_model_domain.

SpectralFitting.make_objective_variance — Method

make_objective_variance(layout::AbstractLayout, dataset::AbstractDataset)

Make the variance vector associated with each objective point.

SpectralFitting.make_output_domain — Method

make_output_domain(layout::AbstractLayout, dataset::AbstractDataset)

Returns the array used as the output domain. That is, in cases where the model input and output map to different domains, the input domain is said to be the model domain, the input domain is said to be the model domain.

The distinction is mainly used for the purposes of simulating data and for visualising data.

SpectralFitting.make_surrogate_harness — Method

make_surrogate_harness(
    model::M,
    lowerbounds::T,
    upperbounds::T;
    optimization_samples = 200,
    seed_samples = 50,
    S::Type = RadialBasis,
    sample_type = SobolSample(),
    verbose = false,
)

Creates and optimizes a surrogate model of type S for model, using wrap_model_as_objective and optimize_accuracy! for optimization_samples iterations. Model is initially seeded with seed_samples points prior to optimization.

Warning

Additive models integrate energies to calculate flux, which surrogate models are currently not capable of. Results for Additive models likely to be inaccurate. This will be patched in a future version.

SpectralFitting.modelkind — Method

modelkind(M::Type{<:AbstractSpectralModel})
modelkind(::AbstractSpectralModel)

Return the kind of model given by M: either Additive, Multiplicative, or Convolutional.

SpectralFitting.numbertype — Method

numbertype(::AbstractSpectralModel)

Get the numerical type of the model. This goes through FitParam, so that the number type returned is as close to a primative as possible.

See also paramtype.

Example

numbertype(PowerLaw()) == Float64

SpectralFitting.objective_transformer — Method

objective_transformer(layout::AbstractLayout, dataset::AbstractDataset)

Used to transform the output of the model onto the output domain. For spectral fitting, this is the method used to do response folding and bin masking.

If none provided, uses the _DEFAULT_TRANSFORMER

function _DEFAULT_TRANSFORMER()
    function _transformer!!(domain, objective)
        objective
    end
    function _transformer!!(output, domain, objective)
        @. output = objective
    end
    _transformer!!
end

SpectralFitting.operation_symbol — Method

operation_symbol(::AbstractCompositeOperator)
operation_symbol(::Type{<:AbstractCompositeOperator})

Obtain the model symbol from a given AbstractCompositeOperator.

SpectralFitting.optimize_accuracy! — Method

optimize_accuracy!(
    surr::AbstractSurrogate,
    obj::Function,
    lb,
    ub;
    sample_type::SamplingAlgorithm = SobolSample(),
    maxiters = 200,
    N_truth = 5000,
    verbose = false,
)

Improve accuracy (faithfullness) of the surrogate model in recreating the objective function.

Samples a new space of N_truth points between lb and ub, and calculates the objective function obj at each. Finds the point with largest MSE between surrogate and objective, and adds the point to the surrogate pool. Repeats maxiters times, adding maxiters points to surrogate model.

Optionally print to stdout the MSE and iteration count with verbose = true.

Note that upper- and lower-bounds should be in the form (E, params...), where E is a single energy and params are the model parameters.

SpectralFitting.paramtype — Method

paramtype(::AbstractSpectralModel)

Get the parameter type of the model. This, unlike numbertype does not go through FitParam.

Example

paramtype(PowerLaw()) == FitParam{Float64}

SpectralFitting.preferred_support — Method

preferred_support(x)

Get the preferred AbstractLayout of x. If multiple supports are available, the DEFAULT_SUPPORT_ORDERING is followed:

DEFAULT_SUPPORT_ORDERING = (ContiguouslyBinned{Nothing}(nothing), OneToOne{Nothing}(nothing))

SpectralFitting.preferred_units — Method

preferred_units(::Type{<:AbstractDataset}, s::AbstractStatistic)
preferred_units(x, s::AbstractStatistic)

Get the preferred units that a given dataset would use to fit the AbstractStatistic in. For example, for ChiSquared, the units of the model may be a rate, however for Cash the preferred units might be counts.

Returning nothing from this function implies there is no unit preference.

If undefined for a derived type, returns nothing.

See also support_units.

SpectralFitting.register_model_data — Method

SpectralFitting.register_model_data(M::Type{<:AbstractSpectralModel}, model_data::ModelDataInfo...)
SpectralFitting.register_model_data(M::Type{<:AbstractSpectralModel}, remote_and_local::Tuple{String,String}...)
SpectralFitting.register_model_data(M::Type{<:AbstractSpectralModel}, filenames::String...)
SpectralFitting.register_model_data(s::Symbol, filenames::String...)

Register filenames as model data associated with the model given by type M or symbol s. This function does not download any files, but rather adds the relevant filenames to a lookup which SpectralFitting.download_model_data consults when invoked, and consequently model data is only downloaded when needed.

Note

It is good practice to use this method immediately after defining a new model with @xspecmodel to register any required datafiles from the HEASoft source code, and therefore keep relevant information together.

Example

# by type
register_model_data(XS_Laor, "ari.mod")
# by symbol
register_model_data(:XS_KyrLine, "KBHline01.fits")

SpectralFitting.response_energy — Method

response_energy(response::ResponseMatrix)

Get the contiguously binned energy corresponding to the input domain of the response matrix. This is equivalent to the model domain.

SpectralFitting.simulate — Method

simulate(model::AbstractSpectralModel; kwargs...)

Returns an InjectiveDataset with a realisation of the model. Can also add noise to the objective, but not to the domain.

The kwargs are:

seed: if not nothing used to set the PRNG seed.

var: variance of the data (assuming mean of 0)

SpectralFitting.support_units — Method

support_units(x)

Return the units of a particular layout. If this method returns nothing, assume the layout does not care about the units and handle that information appropriately (throw an error or set defaults).

SpectralFitting.supports — Method

supports(x::Type)

Used to define whether a given type has support for a specific AbstractLayout. Should return a tuple of the supported layouts. This method should be implemented to express new support, not the query method.

To query, there is

supports(layout::AbstractLayout, x)::Bool

Example

supports(::Type{typeof(x)}) = (OneToOne(),)
@assert supports(ContiguouslyBinned(), x) == false

SpectralFitting.with_units — Method

with_units(::AbstractLayout, units)

Remake the AbstractLayout with the desired units. This may be a no-op if the layout does not care about units, see support_units.

SpectralFitting.wrap_model_as_objective — Method

wrap_model_as_objective(model::AbstractSpectralModel; ΔE = 1e-1)
wrap_model_as_objective(M::Type{<:AbstractSpectralModel}; ΔE = 1e-1)

Wrap a spectral model into an objective function for building/optimizing a surrogate model. Returns an anonymous function taking the tuple (E, params...) as the argument, and returning a single flux value.

SpectralFitting.AbstractModelWrapper — Type

abstract type AbstractModelWrapper{M,T,K} <: AbstractSpectralModel{T,K} end

First field of the struct must be model.

SpectralFitting.AbstractMultiDataset — Type

Must support the same API, but may also have some query methods for specific internals.

SpectralFitting.AbstractSpectralModelImplementation — Type

abstract type AbstractSpectralModelImplementation end

Details about the implementation are represented by this abstract type, used in the trait pattern. Concrete types are

JuliaImplementation(): means the model is implemented in the Julia programming language.

XSPECImplementation(): means the model is vendored from the XSPEC model library.

SpectralFitting.AbstractTableModel — Type

abstract type AbstractTableModel{T,K} <: AbstractSpectralModel{T,K} end

Abstract type representing table models, i.e. those models that interpolate or load data from a table.

First field in the struct must be table. See PhotoelectricAbsorption for an example implementation.

SpectralFitting.XS_CutOffPowerLaw — Type

XS_CutOffPowerLaw(K, Γ, Ecut, z)

K: Normalisation.

Γ: Photon index.

Ecut: Cut-off energy (keV).

z: Redshift.

Example

using SpectralFitting
using UnicodePlots
energy = 10 .^collect(range(-1.0, 2.0, 100))
m = invokemodel(energy, XS_CutOffPowerLaw())
lineplot(energy[1:end-1],m,xscale=:log10,yscale=:log10,xlim=(1e-1,1e2),ylim=(1e-6,1e0),xlabel="Energy (keV)",ylabel="Flux",title="XS_CutOffPowerLaw",canvas=DotCanvas)

                    XS_CutOffPowerLaw             
        ┌────────────────────────────────────────┐ 
10⁰     │:..                                     │ 
        │   '':...                               │ 
        │        '':..                           │ 
        │             '''..                      │ 
        │                  '':..                 │ 
        │                      '':.              │ 
        │                          ':.           │ 
Flux    │                             '..        │ 
        │                               ':.      │ 
        │                                 '.     │ 
        │                                   :    │ 
        │                                    :.  │ 
        │                                     :  │ 
        │                                      : │ 
10⁻⁶    │                                       :│ 
        └────────────────────────────────────────┘ 
         10⁻¹                                 10²  
                        Energy (keV)

SpectralFitting.XS_Jet — Type

XS_Jet(K, mass, Dco, log_mdot, thetaobs, BulkG, phi, zdiss, B, logPrel, gmin_inj, gbreak, gmax, s1, s2, z)

K: MUST BE FIXED AT UNITY as the jet spectrum normalisation is set by the relativisitic particle power.

mass: Black hole mass in solar masses.

Dco: Comoving (proper) distance in Mpc.

log_mdot: log(L/L_Edd.

thetaobs: Inclination angle (deg).

BulkG: Bulk lorentz factor of the jet.

phi: Angular size scale (radians) of the jet acceleration region as seen from the black hole.

zdiss: Vertical distance from the black hole of the jet dissipation region (r_g).

B: Magnetic field in the jet (Gauss).

logPrel: Log of the power injected in relativisitic particles (ergs/s).

gmin_inj: Minimum lorentz factor of the injected electrons.

gbreak: Lorentz factor of the break in injected electron distribution.

gmax: Maximum lorentz factor.

s1: Injected index of the electron distribution below the break.

s2: Injected index of the electron distribution above the break.

z: Cosmological redshift corresponding to the comoving distance used above.

Example

using UnicodePlots
energy = 10 .^collect(range(-8.0, 8.0, 100))
m = invokemodel(energy, XS_Jet())
lineplot(energy[1:end-1],m,xscale=:log10,yscale=:log10,xlim=(1e-8,1e8),ylim=(1e-8,1e8),xlabel="Energy (keV)",ylabel="Flux",title="XS_Jet",canvas=DotCanvas)

                        XS_Jet                   
        ┌────────────────────────────────────────┐ 
10⁸     │                                        │ 
        │                                        │ 
        │                                        │ 
        │                                        │ 
        │  :':..                                 │ 
        │ .'   ''.                               │ 
        │:        ':.                            │ 
Flux    │'          ':                           │ 
        │             '.                         │ 
        │              :                         │ 
        │               ''.....                  │ 
        │                     '''''..            │ 
        │                            ':..        │ 
        │                               ':.      │ 
10⁻⁸    │                                 ''.    │ 
        └────────────────────────────────────────┘ 
        10⁻⁸                                 10⁸  
                        Energy (keV)                

SpectralFitting.XS_Kerrconv — Type

XS_Kerrconv()

Index1: The emissivity index for the inner disk.

Index2: The emissivity index for the outer disk.

r_br_g: The break radius separating the inner and outer portions of the disk, in gravitational radii.

a: The dimensionless spin parameter of the black hole.

Incl: The disk inclination angle, in degrees. A face-on disk has Incl=0.

Rin_ms: The inner radius of the disk, in units of the radius of marginal stability.

Rout_ms: The outer radius of the disk, in units of the radius of marginal stability.

SpectralFitting.XS_Optxagnf — Type

XS_Optxagnf()

K: Normalisation must be frozen.

mass: Black hole mass in solar masses.

Dco: Comoving (proper) distance in Mpc.

logLoLEdd: The Eddington ratio. log(L/L_Edd).

astar: The dimensionless black hole spin.

rcor: The coronal radius in Rg = GM/c^2.

logrout: The log of the outer radius of the disk in units of Rg.

kT_e: The electron temperature for the soft Comptonisation component (soft excess), in keV.

τ: The optical depth of the soft Comptonisation component.

Gamma: The spectral index of the hard Comptonisation component (‘power law’) which has temperature fixed to 100 keV.

fpl: the fraction of the power below rcor which is emitted in the hard comptonisation component.

z: Redshift.

Example

energy = collect(range(0.1, 20.0, 100))
invokemodel(energy, XS_Optxagnf())

                                        XS_Optxagnf                
                        ┌────────────────────────────────────────┐ 
                      0 │                                        │ 
                        │:                                       │ 
                        │:                                       │ 
                        │:                                       │ 
                        │'.                                      │ 
                        │ :                                      │ 
                        │ '.                                     │ 
   Flux (log scale)     │  :.                                    │ 
                        │   :                                    │ 
                        │    '.                                  │ 
                        │     ''':.....                          │ 
                        │              '''''''':.............    │ 
                        │                                    ''''│ 
                        │                                        │ 
                    -30 │                                        │ 
                        └────────────────────────────────────────┘ 
                         0                                     20  
                                          E (keV)               

SpectralFitting.Reflection._add_composite_info! — Method

Used exclusively to do recursive CompositeModel parsing.

SpectralFitting.Reflection.add_objective_reduction! — Method

add_objective_reduction!(ra::CompositeAggregation, op::Symbol)

Reduces the current objective count and applies the reduction operation op to them. For example, if op is :+, and the objective count is 3, then after this function has been called the objective count will be 2 and the reduction expression

@. flux2 = flux2 + flux3

will have been added to the CompositeAggregation.

SpectralFitting.Reflection.assemble_composite_model_call — Method

assemble_composite_model_call(model::Type{<:CompositeModel}, parameters::Type{<:AbstractVector})

Assemble the full composite model call, with objective unpacking via assemble_objective_unpack, closure and parameter assignments, model invocation, and objective reduction. Uses assemble_fast_call to put the final function body together.

SpectralFitting.Reflection.assemble_objective_unpack — Method

assemble_objective_unpack(N)

Assembles the statements for unpacking the objective cache into a number of views. Assembles the part of the model call that looks like:

objective1 = view(objectives, :, 1), objective2 = ...

for N objectives slices.

SpectralFitting.Reflection.get_info — Method

get_info(model::Type{<:AbstractSpectralModel}, lens::Lens)

Returns a ModelInfo struct for a given model.

SpectralFitting.Reflection.make_constructor — Method

make_constructor(model::Type{<:AbstractSpectralModel}, closures::Vector, params::Vector, T::Type)

Create a constructor expression for the model. Should return something similar to

:(PowerLaw{T}(arg1, arg2, arg3)))

unpacking the closures and params vectors in the appropriate places.

SpectralFitting.Reflection.parameter_lenses — Method

parameter_lenses(::Type{<:AbstractSpectralModel}, info::ModelInfo)

Return a vector of lenses Lens that refer to each of the models parameters. The lenses should be relative to model.lens.

SpectralFitting.Reflection.set_objective_count! — Method

set_objective_count!(a::CompositeAggregation, o::Int)

Set the objective count to a specific value. Will bump the maximum objective count if the new value exceeds the current maximum.

SpectralFitting.Reflection.CompositeModelInfo — Type

struct CompositeModelInfo
    "The parameter symbols of the model with the respective lens to the actual parameter."
    parameter_symbols::Vector{Pair{Symbol,Lens}}
    "Each model assigned to a unique symbol."
    models::Vector{Pair{Symbol,ModelInfo}}
    "The expression representing the folding operations of this composite model."
    model_expression::Expr
    "Constructor and objective folding expressions, used in generating the invocation call."
    expressions::Vector{Expr}
    "The maximum number of objective caches this model will need."
    maximum_objective_cache_count::Int
    "How many objective caches are currently active."
    objective_cache_count::Int
end

The composite equivalent of ModelInfo, augmented to track the model symbol (a1, m3, etc.), and the model parameters (K_1, a_3, etc.)

SpectralFitting.Reflection.ModelInfo — Type

struct ModelInfo
    "All parameter symbols for the model."
    symbols::Vector{Symbol}
    "Unique symbols generated for the parameter assignment when buildin the function call."
    generated_symbols::Vector{Symbol}
    "Additional closure parameters that need to be handled when invoking the model."
    closure_symbols::Vector{Symbol}
    "Unique closure generated symbols."
    generated_closure_symbols::Vector{Symbol}
    "The lens that takes you to this model from some parent."
    lens::Lens
    "The model type itself."
    model::Type
end

All models are parsed into a ModelInfo struct relative to their parent (in the case of composite models).

The symbols field contains all of the model parameter symbols as they are in the structure, not as they have been generated. Recall when the invocation expressions are generated, we create anonymous paramter names to avoid conflicts. These are the generated_symbols instead.