API reference

Energy and flux utility functions

Missing docstring.

Missing docstring for make_flux. Check Documenter's build log for details.

Missing docstring.

Missing docstring for make_fluxes. Check Documenter's build log for details.

Utility functions for defining new models

SpectralFitting.finite_diff_kernel! — Function

finite_diff_kernel!(f::Function, flux, energy)

Calculates the finite difference of the function f over the energy bin between the high and low bin edges, via

\[c_i = f(E_{i,\text{high}}) - f(E_{i,\text{low}}),\]

similar to evaluating the limits of the integral between $E_{i,\text{high}}$ and $E_{i,\text{low}}$.

This utility function is primarily used for Additive models to ensure the flux per bin is normalised for the energy over the bin.

SpectralFitting.wrap_model_as_objective — Function

wrap_model_as_objective(model::AbstractSpectralModel; ΔE = 1e-1)
wrap_model_as_objective(M::Type{<:AbstractSpectralModel}; ΔE = 1e-1)

Wrap a spectral model into an objective function for building/optimizing a surrogate model. Returns an anonymous function taking the tuple (E, params...) as the argument, and returning a single flux value.

Introspection

Missing docstring.

Missing docstring for has_closure_params. Check Documenter's build log for details.

Missing docstring.

Missing docstring for get_closure_param_fields. Check Documenter's build log for details.

Missing docstring.

Missing docstring for model_base_name. Check Documenter's build log for details.