Surrogate models

Surrogate models allow you to create fast or memory efficient approximations of model components, or assist in optimizing some objective function directly. SpectralFitting uses the Surrogates.jl library of models, that yields pure-Julia surrogate models. Consequently, surrogate models also permit use of automatic differentiation in fitting, and are therefore powerful tools for improving fitting performance.

Surrogates overview

Any function may be wrapped as a surrogate model using the SurrogateSpectralModel type.

SpectralFitting.SurrogateSpectralModel — Type

SurrogateSpectralModel <: AbstractSpectralModel
SurrogateSpectralModel(modelkind, surrogate, params, params_symbols)

Used to wrap a surrogate function into an AbstractSpectralModel.

Example

Creating a surrogate function using make_surrogate_function:

# build and optimize a surrogate model
surrogate = make_surrogate_function(model, lower_bounds, upper_bounds)

# create surrogate spectral model
sm = SurrogateSpectralModel(
    Multiplicative(),
    surrogate,
    (FitParam(1.0),),
    (:ηH,)
)

The lower_bounds and upper_bounds must be tuples in the form (E, params...), where E denotes the bounds on the energy range to train over.

To facilitate easy surrogate builds, SpectralFitting exports a number of utility functions.

SpectralFitting.make_surrogate_function — Function

make_surrogate_function(
    model::M,
    lowerbounds::T,
    upperbounds::T;
    optimization_samples = 200,
    seed_samples = 50,
    S::Type = RadialBasis,
    sample_type = SobolSample(),
    verbose = false,
)

Creates and optimizes a surrogate model of type S for model, using wrap_model_as_objective and optimize_accuracy! for optimization_samples iterations. Model is initially seeded with seed_samples points prior to optimization.

Warning

Additive models integrate energies to calculate flux, which surrogate models are currently not capable of. Results for Additive models likely to be inaccurate. This will be patched in a future version.

SpectralFitting.optimize_accuracy! — Function

optimize_accuracy!(
    surr::AbstractSurrogate,
    obj::Function,
    lb,
    ub;
    sample_type::SamplingAlgorithm = SobolSample(),
    maxiters = 200,
    N_truth = 5000,
    verbose = false,
)

Improve accuracy (faithfullness) of the surrogate model in recreating the objective function.

Samples a new space of N_truth points between lb and ub, and calculates the objective function obj at each. Finds the point with largest MSE between surrogate and objective, and adds the point to the surrogate pool. Repeats maxiters times, adding maxiters points to surrogate model.

Optionally print to stdout the MSE and iteration count with verbose = true.

Note that upper- and lower-bounds should be in the form (E, params...), where E is a single energy and params are the model parameters.

Creating a surrogate for XS_PhotoelectricAbsorption

Before we start, let us discuss a number of benefits the use of surrogate models may bring us:

SurrogateSpectralModel permit use of automatic differentiation.

Surrogate models may be allocation-free depending on setup, whereas XSPEC wrappers will always have to allocate for type-conversions.

Surrogate models may be considerably faster, especially for table models.

Surrogate models are shareable (see Sharing surrogate models), and are tunable in size.

XS_PhotoelectricAbsorption is an XSPEC model that is wrapped by a thin C-wrapper into Julia. The implementation of this model is a number of Fortran routines from the late 90s, including a tabulation of ~3000 lines of data that has been copied directly into the Fortran source code.

The performance of this model represents its complexity.

energy = collect(range(0.1, 20.0, 100))
flux = make_flux(energy)

model = XS_PhotoelectricAbsorption()

XS_PhotoelectricAbsorption
   ηH  => (1 ± 0.1)

Benchmarking with BenchmarkTools.jl:

using BenchmarkTools
@btime invokemodel!($flux, $energy, $model);

  2.302 ms (3 allocations: 112 bytes)

The surrogate we'll construct will have to be tailored a little to the data we wish to fit, as we need to specify the parameter ranges our surrogate should learn. For example, we might be interested in energies between $0.1$ and $20$ keV, with equivalent hydrogen column $\eta$H anywhere between $10^{-3}$ and $30$. We specify these bounds using tuples

lower_bounds = (0.1, 1e-3)
upper_bounds = (20.0, 30.0)

Note

The first index is always the energy bounds, and the subsequent indices are the parameters in the same order they are defined in the model structure.

Next, we use make_surrogate_function to build and optimize a surrogate function for our model. By default, the surrogate uses linear radial basis functions, and seeds the coefficients with a number of seed points. This function then improves the accuracy of the model using optimize_accuracy!, until a maximal number of iterations has been reached.

For illustration purposes, we'll omit the accuracy improving step, and perform this ourselves. We can do this by setting optimization_samples = 0 in the keyword arguments:

surrogate = make_surrogate_function(
    model,
    lower_bounds,
    upper_bounds
    ;
    optimization_samples = 0
)

# number of points the surrogate has been trained on
length(surrogate.x)

50

We can examine how well our surrogate reconstructs the model for a given test parameter:

# random test value
ηh_test = rand(1.0:30.0)

f = invokemodel(energy, model, ηh_test)

f̂ = map(energy[1:end-1]) do e
    v = (e, ηh_test)
    surrogate(v)
end

Now we'll use optimize_accuracy! to improve the faithfulness of our surrogate. This requires making use of wrap_model_as_objective as a little wrapper around our model:

obj = wrap_model_as_objective(model)

optimize_accuracy!(surrogate, obj, lower_bounds, upper_bounds; maxiters=250)

length(surrogate.x)

300

We can plot the surrogate model again and see the improvement.

new_f̂ = map(energy[2:end]) do e
    v = (e, ηh_test)
    surrogate(v)
end

Tight. We can also inspect the memory footprint of our model:

# in bytes
Base.summarysize(surrogate)

9808

This may be reduced by lowering maxiters in optimize_accuracy! at the cost of decreasing faithfulness. However, compare this to the Fortran tabulated source file in the XSPEC source code, which is approximately 224 Kb – about 22x larger. The surrogate models are considerably more portable at this level.

Inspecting the plot shows that there is a small domain near zero energy where the surrogate is calculating a negative value. We can either try to improve the accuracy to fix this, or specify a clamping function: for XS_PhotoelectricAbsorption, we know the multiplicative factor is constrained between 0 and 1. Therefore, we may adjust our surrogate to reflect this:

clamped_surrogate = (v) -> clamp(surrogate(v), 0.0, 1.0)

Using a surrogate spectral model

Now that we have the surrogate model, we use SurrogateSpectralModel to wrap it into an AbstractSpectralModel. The constructor also needs to know the model kind, have a copy of the model parameters, and know which symbols to represent the parameters with.

sm = SurrogateSpectralModel(
    Multiplicative(),
    clamped_surrogate,
    (FitParam(1.0),), # must be a tuple
    (:ηh,) # must also be a tuple
)

SurrogateSpectralModel{Multiplicative}
   ηh  => (1 ± 0.1)

We can now use the familiar API and attempt to benchmark the performance:

@btime invokemodel!($flux, $energy, $sm);

  50.406 μs (297 allocations: 6.19 KiB)

These allocations are coming from the global closure we have in the clamped_surrogate lambda. We can actually elide these using a proper closure:

make_surr(surrogate) = SpectralFitting.SurrogateSpectralModel(
    Multiplicative(),
    (v) -> clamp(surrogate(v), 0.0, 1.0),
    (FitParam(1.0),),
    (:ηH,)
)

sm2 = make_surr(surrogate)

SurrogateSpectralModel{Multiplicative}
   ηH  => (1 ± 0.1)

Now when we benchmark

@btime invokemodel!($flux, $energy, $sm2);

  46.228 μs (0 allocations: 0 bytes)

Comparing this to the initial benchmark of XS_PhotoelectricAbsorption, we see about a 60x speedup, with no allocations, and this surrogate model is now automatic differentiation ready.

Sharing surrogate models

To export and import surrogate models, JLD2.jl is recommended.