Using spectral models

Models exist as three different kinds, defined by an AbstractSpectralModelKind trait.

SpectralFitting.AbstractSpectralModelKind — Type

abstract type AbstractSpectralModelKind

Abstract type of all model kinds. The algebra of models is as follows

A + A = A
M * M = M
M * A = A
C(A)  = A

where A is Additive, M is Multiplicative, and C is Convolutional. All other operations are prohibited, e.g. C(M) or M * C. To obtain M * C there must be an additive component, e.g. M * C(A).

SpectralFitting.Additive — Type

Additive <: AbstractSpectralModelKind
Additive()

Additive models are effectively the sources of photons, and are the principle building blocks of composite models. Every additive model has a normalisation parameter which re-scales the flux by a constant factor K.

Note

Defining custom additive models requires special care. See Defining new models.

SpectralFitting.Multiplicative — Type

Multiplicative <: AbstractSpectralModelKind
Multiplicative()

Multiplicative models act on Additive models, by element-wise multiplying the flux in each energy bin of the additive model by a different factor.

SpectralFitting.Convolutional — Type

Convolutional <: AbstractSpectralModelKind
Convolutional()

Convolutional models act on the flux generated by Additive models, similar to Multiplicative models, however may convolve kernels through the flux also.

Model abstraction

All spectral models are a sub-type of AbstractSpectralModel, including CompositeModel and SurrogateSpectralModel.

SpectralFitting.AbstractSpectralModel — Type

abstract type AbstractSpectralModel

Supertype of all spectral models. Sub-types must implement the following interface

modelkind

SpectralFitting.invoke!

SpectralFitting.invoke! — Function

SpectralFitting.invoke!(flux, energy, M::Type{<:AbstractSpectralModel}, params...)

Used to define the behaviour of models. Should calculate flux of the model and write in-place into flux.

Warning

This function should not be called directly. Use invokemodel instead.

Parameters are passed in in-order as defined in the model structure. For example

struct MyModel{F1,F2,F3,...} <: AbstractSpectralModel
    p1::F1
    p2::F2
    p3::F3
    # ...
end

would have the arguments passed to invoke! as

function SpectralFitting.invoke!(flux, energy, ::Type{<:MyModel}, p1, p2, p3, ...)
    # ...
end

The only exception to this are Additive models, where the normalisation parameter K is not passed to invoke!.

SpectralFitting.modelkind — Function

modelkind(M::Type{<:AbstractSpectralModel})

Return the kind of model given by M: either Additive, Multiplicative, or Convolutional.

Model data availability

Many of the XSPEC implemented models use tabular data, such as FITS, and return results interpolated from these pre-calculated tables. In some cases, these table models have data files that are multiple gigabytes in size, and would be very unwieldy to ship indiscriminantly. SpectralFitting attempts to circumnavigate this bloat by downloading the model data on an ut opus basis.

SpectralFitting.download_model_data — Function

SpectralFitting.download_model_data(model::AbstractSpectralModel; kwargs...)
SpectralFitting.download_model_data(M::Type{<:AbstractSpectralModel}; kwargs...)
SpectralFitting.download_model_data(s::Symbol; kwargs...)

Downloads the model data for a model specified either by model, type M, or symbol s. Datafiles associated with a specific model may be registered using SpectralFitting.register_model_data. The download is currently unconfigurable, but permits slight control via a number of keyword arguments:

progress::Bool = true

Display a progress bar for the download.

model_source_url::String = "http://www.star.bris.ac.uk/fbaker/XSPEC-model-data"

The source URL used to download the model data.

All standard XSPEC spectral model data is currently being hosted on the University of Bristol astrophysics servers, and should be persistently available to anyone.

SpectralFitting.download_all_model_data — Function

SpectralFitting.download_all_model_data()

Downloads all model data for the models currently registered with SpectralFitting.register_model_data. Calls SpectralFitting.download_model_data to perform the download.

Special care must be taken if new XSPEC models are wrapped to ensure the data is available. For more on this, see Wrapping new XSPEC models.

Model data may also alternatively be copied in by-hand from a HEASoft XSPEC source directory. In this case, the location to copy the data to may be determined via joinpath(SpectralFitting.LibXSPEC_jll.artifact_dir, "spectral", "modelData").

Instantiating and invoking models

Models may be composed together to create more complex spectra, with the algebra defined by the AbstractSpectralModelKind.

For example, a XS_PowerLaw model may multiply element-wise by a XS_PhotoelectricAbsorption model:

model = XS_PhotoelectricAbsorption() * XS_PowerLaw()
model

CompositeModel with 2 component models:
     m1 * a1
 Symbols:    Models:                     Params:
       a1 => XS_PowerLaw                 (K_1, a_1) 
       m1 => XS_PhotoelectricAbsorption  (ηH_1) 
 Symbols:     Values:
      K_1 =>    1 ± 0.1   ∈ [ 0, Inf ]   FREE
      a_1 =>  0.5 ± 0.05  ∈ [ 0, Inf ]   FREE
     ηH_1 =>    1 ± 0.1   ∈ [ 0, Inf ]   FREE

A model may be evaluated by using invokemodel:

energy = collect(range(0.1, 20.0, 100))
invokemodel(energy, model)

99-element Vector{Float64}:
 0.0
 7.613847449558353e-9
 0.0007743430511093221
 0.003860477072949348
 0.01504614049721437
 0.037908600300382914
 0.059153540665640825
 0.07462294243145176
 0.08495775522244775
 0.09110310891499386
 ⋮
 0.046901156794963
 0.04664859998627752
 0.046400022418344525
 0.046155319337826746
 0.04591439540016139
 0.045677153431287774
 0.04544350525190907
 0.04521336057578862
 0.04498663773824884

Note

XSPEC models are implemented to accept a single energy vector with all of the low and high energy bin edges, and return a flux array with the flux in each energy bin. For consistency, SpectralFitting implement models in the same way, and therefore it is always the case that:

length(flux) == length(energy) - 1

SpectralFitting.invokemodel — Function

invokemodel(energy, model)
invokemodel(energy, model, free_params)

Invoke the AbstractSpectralModel given by model, optionally overriding the free parameters with values given in free_params. free_params may be a vector or tuple with element type AbstractFitParameter or Number.

This function, unlike SpectralFitting.invoke! used to define models, is sensitive to performing any normalisation or post-processing tasks that a specific model kind may require.

Note

Users should always call models using invokemodel or invokemodel! to ensure normalisations and closures are accounted for.

invokemodel allocates the needed flux arrays based on the element type of free_params to allow automatic differentation libraries to calculate parameter gradients.

In-place non-allocating variants are the invokemodel! functions.

Example

model = XS_PowerLaw()
energy = collect(range(0.1, 20.0, 100))
invokemodel(energy, model)

p0 = [0.1, 2.0] # change K and a
invokemodel(energy, model, p0)

SpectralFitting.invokemodel! — Function

invokemodel!(flux, energy, model)
invokemodel!(flux, energy, model, free_params)
invokemodel!(flux, energy, model, free_params, frozen_params)

In-place variant of invokemodel, calculating the flux of an AbstractSpectralModel given by model, optionally overriding the free and/or frozen parameter values. These arguments may be a vector or tuple with element type AbstractFitParameter or Number.

The number of fluxes to allocate for a model may change if using any CompositeModel as the model. It is generally recommended to use flux_count to ensure the correct number of flux arrays are allocated with make_fluxes when using composite models.

Single spectral model components should use make_flux instead.

Example

model = XS_PowerLaw()
energy = collect(range(0.1, 20.0, 100))
flux = make_flux(energy)
invokemodel!(flux, energy, model)

p0 = [0.1, 2.0] # change K and a
invokemodel!(flux, energy, model, p0)

invokemodel!(flux, energy, ::Type{<:AbstractSpectralModel}, params...)

Specialisation of invokemodel! which dispatches on types. This is primarily used in the generated functions, to assemble efficient model calls at compile time. In general, users should prefer one of the other dispatches.

Utilities

SpectralFitting defines a number of utility functions for model introspection and manipulation.

SpectralFitting.flux_count — Function

flux_count(model::AbstractSpectralModel)

Returns the number of flux arrays the model needs when using invokemodel!.

Example

model = XS_PhotoelectricAbsorption() * XS_PowerLaw()
flux_count(model)

SpectralFitting.update_params! — Function

update_params!(model::AbstractSpectralModel, values)
update_params!(model::AbstractSpectralModel, values, errors)

Update the free model parameters with values, optionally also updating the errors.

SpectralFitting.get_param — Function

get_param(model::AbstractSpectralModel, s::Symbol)

Get a parameter from an AbstractSpectralModel by symbol.

Example

model = XS_BlackBody()
get_param(model, :K)

SpectralFitting.get_param_count — Function

get_param_count(model::AbstractSpectralModel)
get_param_count(M::Type{<:AbstractSpectralModel})

Get the number of parameters a given AbstractSpectralModel has.

Example

model = XS_BlackBody() + XS_PowerLaw()
get_param_count(model)

SpectralFitting.get_param_types — Function

get_param_types(model::AbstractSpectralModel)
get_param_types(M::Type{<:AbstractSpectralModel})

Get a Vector{Type} with the type of each parameters in a given AbstractSpectralModel.

Example

model = XS_BlackBody() + XS_PhotoelectricAbsorption() * XS_PowerLaw()
get_param_types(model)

SpectralFitting.get_param_symbols — Function

get_param_symbols(model::AbstractSpectralModel)
get_param_symbols(M::Type{<:AbstractSpectralModel})

Get a Vector{Symbol} with the symbol of each parameters in a given AbstractSpectralModel.

Example

model = XS_PhotoelectricAbsorption() * XS_Laor()
get_param_symbols(model)

SpectralFitting.get_param_symbol_pairs — Function

get_param_symbol_pairs(m::AbstractSpectralModel)

Get a generator yielding ::Pair{Symbol,T} of all model parameter values and their symbols, for an AbstractSpectralModel.

Example

model = XS_BlackBody() + XS_PowerLaw()
get_param_symbol_pairs(model)

SpectralFitting.get_params — Function

get_params(m::AbstractSpectralModel)

Get a generator of all model parameters of an AbstractSpectralModel.

Example

model = XS_BlackBody() + XS_PowerLaw()
get_params(model)

SpectralFitting.get_params_value — Function

get_params_value(m::AbstractSpectralModel)

Get a generator of all model parameter values of an AbstractSpectralModel. See get_value.

Example

model = XS_BlackBody() + XS_PowerLaw()
get_params_value(model)