Using spectral models

Models exist as three different kinds, defined by an AbstractSpectralModelKind trait.

SpectralFitting.AbstractSpectralModelKind — Type

abstract type AbstractSpectralModelKind

Abstract type of all model kinds. The algebra of models is as follows

A + A = A
M * M = M
M * A = A
C(A)  = A

where A is Additive, M is Multiplicative, and C is Convolutional. All other operations are prohibited, e.g. C(M) or M * C. To obtain M * C there must be an additive component, e.g. M * C(A).

SpectralFitting.Additive — Type

Additive <: AbstractSpectralModelKind
Additive()

Additive models are effectively the sources of photons, and are the principle building blocks of composite models. Every additive model has a normalisation parameter which re-scales the output by a constant factor K.

Note

Defining custom additive models requires special care. See Defining new models.

SpectralFitting.Multiplicative — Type

Multiplicative <: AbstractSpectralModelKind
Multiplicative()

Multiplicative models act on Additive models, by element-wise multiplying the output in each energy bin of the additive model by a different factor.

SpectralFitting.Convolutional — Type

Convolutional <: AbstractSpectralModelKind
Convolutional()

Convolutional models act on the output generated by Additive models, similar to Multiplicative models, however may convolve kernels through the output also.

Model abstraction

All spectral models are a sub-type of AbstractSpectralModel, including CompositeModel and SurrogateSpectralModel.

SpectralFitting.AbstractSpectralModel — Type

abstract type AbstractSpectralModel{T,K}

Supertype of all spectral models. Sub-types must implement the following interface

modelkind

SpectralFitting.invoke!

The parametric type parameter T is the number type of the model and K defines the AbstractSpectralModelKind.

SpectralFitting.invoke! — Function

SpectralFitting.invoke!(output, energy, M::Type{<:AbstractSpectralModel}, params...)

Used to define the behaviour of models. Should calculate output of the model and write in-place into output.

Warning

This function should not be called directly. Use invokemodel instead.

Parameters are passed in in-order as defined in the model structure. For example

struct MyModel{F1,F2,F3,...} <: AbstractSpectralModel
    p1::F1
    p2::F2
    p3::F3
    # ...
end

would have the arguments passed to invoke! as

function SpectralFitting.invoke!(output, energy, ::Type{<:MyModel}, p1, p2, p3, ...)
    # ...
end

The only exception to this are Additive models, where the normalisation parameter K is not passed to invoke!.

SpectralFitting.modelkind — Function

modelkind(M::Type{<:AbstractSpectralModel})
modelkind(::AbstractSpectralModel)

Return the kind of model given by M: either Additive, Multiplicative, or Convolutional.

Model data availability

Many of the XSPEC implemented models use tabular data, such as FITS, and return results interpolated from these pre-calculated tables. In some cases, these table models have data files that are multiple gigabytes in size, and would be very unwieldy to ship indiscriminantly. SpectralFitting attempts to circumnavigate this bloat by downloading the model data on an ut opus basis.

SpectralFitting.download_model_data — Function

SpectralFitting.download_model_data(model::AbstractSpectralModel; kwargs...)
SpectralFitting.download_model_data(M::Type{<:AbstractSpectralModel}; kwargs...)
SpectralFitting.download_model_data(s::Symbol; kwargs...)

Downloads the model data for a model specified either by model, type M, or symbol s. Datafiles associated with a specific model may be registered using SpectralFitting.register_model_data. The download is currently unconfigurable, but permits slight control via a number of keyword arguments:

progress::Bool = true

Display a progress bar for the download.

model_source_url::String = "http://www.star.bris.ac.uk/fbaker/XSPEC-model-data"

The source URL used to download the model data.

All standard XSPEC spectral model data is currently being hosted on the University of Bristol astrophysics servers, and should be persistently available to anyone.

SpectralFitting.download_all_model_data — Function

SpectralFitting.download_all_model_data()

Downloads all model data for the models currently registered with SpectralFitting.register_model_data. Calls SpectralFitting.download_model_data to perform the download.

Special care must be taken if new XSPEC models are wrapped to ensure the data is available. For more on this, see Wrapping new XSPEC models.

Model data may also alternatively be copied in by-hand from a HEASoft XSPEC source directory. In this case, the location to copy the data to may be determined via joinpath(SpectralFitting.LibXSPEC_jll.artifact_dir, "spectral", "modelData").

Instantiating and invoking models

Models may be composed together to create more complex spectra, with the algebra defined by the AbstractSpectralModelKind.

For example, a XS_PowerLaw model may multiply element-wise by a XS_PhotoelectricAbsorption model:

model = XS_PhotoelectricAbsorption() * XS_PowerLaw()
model

┌ CompositeModel with 2 component models:
│      m1 * a1
│ Model key and parameters:
│    a1 => XS_PowerLaw
│       K_1 ->  1 ± 0.1  ∈ [ 0, Inf ]   FREE
│       a_1 ->  1 ± 0.1  ∈ [ 0, Inf ]   FREE
│    m1 => XS_PhotoelectricAbsorption
│      ηH_1 ->  1 ± 0.1  ∈ [ 0, Inf ]   FREE
└ 

A model may be evaluated by using invokemodel:

energy = collect(range(0.1, 20.0, 100))
invokemodel(energy, model)

99-element view(::Matrix{Float64}, :, 1) with eltype Float64:
 0.0
 1.2178552349564805e-8
 0.0010034521654331123
 0.004320802901695416
 0.015043548132996451
 0.03457602289338916
 0.04993025710630304
 0.05890347901257127
 0.06321400078471338
 0.06429917452380776
 ⋮
 0.010966363689722607
 0.010847867271882167
 0.010731891106496139
 0.010618355388832514
 0.010507184929488832
 0.010398306430377402
 0.010291650845230853
 0.01018715072973584
 0.010084742568964479

Note

XSPEC models are implemented to accept a single energy vector with all of the low and high energy bin edges, and return a flux array with the flux in each energy bin. For consistency, SpectralFitting implement models in the same way, and therefore it is always the case that:

length(flux) == length(energy) - 1

SpectralFitting.invokemodel — Function

invokemodel(energy, model)
invokemodel(energy, model, free_params)

Invoke the AbstractSpectralModel given by model, optionally overriding the free parameters with values given in free_params. free_params may be a vector or tuple with element type FitParam or Number.

This function, unlike SpectralFitting.invoke! used to define models, is sensitive to performing any normalisation or post-processing tasks that a specific model kind may require.

Note

Users should always call models using invokemodel or invokemodel! to ensure normalisations and closures are accounted for.

invokemodel allocates the needed output arrays based on the element type of free_params to allow automatic differentation libraries to calculate parameter gradients.

In-place non-allocating variants are the invokemodel! functions.

Example

model = XS_PowerLaw()
energy = collect(range(0.1, 20.0, 100))
invokemodel(energy, model)

p0 = [0.1, 2.0] # change K and a
invokemodel(energy, model, p0)

Missing docstring.

Missing docstring for invokemodel!. Check Documenter's build log for details.

Utilities

SpectralFitting defines a number of utility functions for model introspection and manipulation.

Missing docstring.

Missing docstring for flux_count. Check Documenter's build log for details.

Missing docstring.

Missing docstring for get_param. Check Documenter's build log for details.

Missing docstring.

Missing docstring for get_param_count. Check Documenter's build log for details.

Missing docstring.

Missing docstring for get_param_types. Check Documenter's build log for details.

Missing docstring.

Missing docstring for get_param_symbols. Check Documenter's build log for details.

Missing docstring.

Missing docstring for get_param_symbol_pairs. Check Documenter's build log for details.

Missing docstring.

Missing docstring for get_params. Check Documenter's build log for details.

Missing docstring.

Missing docstring for get_params_value. Check Documenter's build log for details.